% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{angle-conversion}
\alias{angle-conversion}
\alias{rad2deg}
\alias{deg2rad}
\title{Degrees to Radians}
\usage{
rad2deg(rad)

deg2rad(deg)
}
\arguments{
\item{rad}{(array of) angles in radians.}

\item{deg}{(array of) angles in degrees.}
}
\value{
The angle in degrees or radians.
}
\description{
Transforms between angles in degrees and radians.
}
\examples{
deg2rad(c(0, 10, 20, 30, 40, 50, 60, 70, 80, 90))
rad2deg(seq(-pi / 2, pi / 2, length = 19))
}
