% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{angle_vectors}
\alias{angle_vectors}
\title{Angle between two vectors}
\usage{
angle_vectors(x, y)
}
\arguments{
\item{x, y}{Vectors in Cartesian coordinates. Can be vectors of three numbers
or a matrix of 3 columns (x, y, z)}
}
\value{
numeric; angle in degrees
}
\description{
Calculates the angle between two vectors
}
\examples{
u <- c(1, -2, 3)
v <- c(-2, 1, 1)
angle_vectors(u, v)
}
