% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{axes}
\alias{axes}
\title{Plot axes}
\usage{
axes(
  x,
  y,
  angle,
  radius = 0.5,
  arrow.code = 1,
  arrow.length = 0,
  add = FALSE,
  ...
)
}
\arguments{
\item{x, y}{coordinates of points}

\item{angle}{Azimuth in degrees}

\item{radius}{length of axis}

\item{arrow.code}{integer. Kind of arrow head. The default is \code{1}, i.e. no arrow head. See \code{\link[graphics:arrows]{graphics::arrows()}} for details}

\item{arrow.length}{numeric Length of the edges of the arrow head (in inches). (Ignored if \code{arrow.code = 1})}

\item{add}{logical. add to existing plot?}

\item{...}{optional arguments passed to \code{\link[graphics:arrows]{graphics::arrows()}}}
}
\value{
No return value, called for side effects
}
\description{
Plot axes
}
\examples{
data("san_andreas")
axes(san_andreas$lon, san_andreas$lat, san_andreas$azi, add = FALSE)
}
