% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{data}
\name{borders}
\alias{borders}
\title{#' Plate Boundaries on the Earth (GSRM 2.1)
#'
#' Global set of present plate boundaries on the Earth based on GSRM2.1 model by
#' Kreemer et al (2014).
#'
#' @docType data
#'
#' @usage data('gsrm2_plates')
#'
#' @format An object of class \code{sf}
#'
#' @references Kreemer, C., Blewitt, G., & Klein, E. C. (2014). A geodetic plate
#' motion and Global Strain Rate Model. \emph{Geochemistry, Geophysics, Geosystems},
#' \strong{15}(10), 3849<U+2013>3889. doi: 10.1002/2014GC005407.
#'
#' @keywords datasets
#'
#' @examples
#' data("gsrm2_plates")
#' head("gsrm2_plates")
"gsrm2_plates"
Borders}
\format{
An object of class \code{sf}
}
\usage{
data('borders')
}
\description{
Global political borders from \code{rnaturalearth}
}
\examples{
data("borders")
head("borders")
}
\keyword{datasets}
