% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{cartesian_to_geographical}
\alias{cartesian_to_geographical}
\title{Cartesian to geographical coordinates}
\usage{
cartesian_to_geographical(n)
}
\arguments{
\item{n}{Vector of three numbers (x, y, z)}
}
\value{
vector of two numbers (latitude, longitude)
}
\description{
Converts vector from Cartesian into geographical coordinates
}
\examples{
u <- c(1, -2, 3)
cartesian_to_geographical(u)
}
\seealso{
\code{\link{geographical_to_cartesian}}
}
