% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circular_dispersion}
\alias{circular_dispersion}
\title{Circular dispersion}
\usage{
circular_dispersion(x, from = NULL, w = NULL, axial = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{Data values. A vector of numeric values in degrees}

\item{from}{numeric. the angle (in degree) for which the dispersion should be measured.
If NULL, the circular median is used.}

\item{w}{(optional) Weights. A vector of positive numbers, of the same length as
\code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(TRUE, the default) or circular, i.e. 2pi-periodical (FALSE).}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}
}
\description{
Dispersion of angles about a given angle.
}
\examples{
\dontrun{
data("san_andreas")
circular_dispersion(san_andreas$azi)

data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, ep, "right")
circular_dispersion(sa.por$azi.PoR, from = 135)
}
}
