% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circular_distance}
\alias{circular_distance}
\title{Circular distance between two angles}
\usage{
circular_distance(x, y, axial = TRUE, na.rm = TRUE)
}
\arguments{
\item{x, y}{vector of numeric values in degrees}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(TRUE, the default) or circular, i.e. 2pi-periodical (FALSE).}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}
}
\description{
Circular distance between two angles
}
