% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circular_mode}
\alias{circular_mode}
\title{Circular Mode}
\usage{
circular_mode(x, ...)
}
\arguments{
\item{x}{numeric. A vector of angles (in degrees) from which the estimate is to be computed.}

\item{...}{parameters passed to \code{\link[=circular_density]{circular_density()}}}
}
\value{
numeric
}
\description{
MLE angle (maximum density) using a circular distribution kernel with
specified concentration
}
\examples{
set.seed(20250411)
x <- rvm(10, 0, 100)

# Mode of von Mises kernel density (the default)
circular_mode(x)

# Mode of wrapped Cauchy kernel density
circular_mode(x, kernel = "wrappedcauchy")
}
\references{
N.I. Fisher (1993) Statistical Analysis of Circular Data, Cambridge University Press.
}
