% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{circular_plot}
\alias{circular_plot}
\title{Circular plot}
\usage{
circular_plot(
  main = NULL,
  labels = TRUE,
  at = seq(0, 360 - 45, 45),
  cborder = TRUE,
  ...
)
}
\arguments{
\item{main}{Character string specifying the title of the plot.}

\item{labels}{Either a logical value indicating whether to plot labels
next to the tick marks, or a vector of labels for the tick marks.}

\item{at}{Optional vector of angles at which tick marks should be plotted.
Set \code{at=numeric(0)} to suppress tick marks.}

\item{cborder}{logical. Border of rose plot.}

\item{...}{optional arguments passed to \code{plot.default()}}
}
\value{
none
}
\description{
Circular plot
}
\note{
Polar diagram where angles increase clockwise.
}
\keyword{internal}
