% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circular_quasi_median}
\alias{circular_quasi_median}
\title{Quasi Median on a Circle}
\usage{
circular_quasi_median(x)
}
\arguments{
\item{x}{a numeric vector containing the orientations whose median is to be
computed}
}
\description{
Median of orientation data, i.e. pi-periodical data
}
\examples{
circular_quasi_median(x = c(0, 45, 55, 40 + 180, 50 + 180))
}
\references{
Ratanaruamkarn, S., Niewiadomska-Bugaj, M., Wang, J.-C. (2009).
A New Estimator of a Circular Median. Communications in Statistics -
Simulation and Computation, 38(6), 1269–1291.
https://doi.org/10.1080/03610910902899950
}
\seealso{
\code{\link[stats]{median}}
}
