% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_chi2.R
\name{circular_quasi_quartile}
\alias{circular_quasi_quartile}
\title{Quasi Quartile on a Circle}
\usage{
circular_quasi_quartile(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector containing the orientations whose sample quartile is to be computed}

\item{na.rm}{logical; if true (default), any NA and NaN's are removed from x before the quantiles are computed.}
}
\description{
The \code{"stats::quantile"} equivalent for circular orienation data
}
\examples{
circular_quasi_quartile(x = c(0, 45, 55, 40+180, 50+180))
}
\references{
Ratanaruamkarn, S., Niewiadomska-Bugaj, M., Wang, J.-C. (2009). A New Estimator of a Circular Median. Communications in Statistics - Simulation and Computation, 38(6), 1269–1291. https://doi.org/10.1080/03610910902899950
}
\seealso{
\code{\link[stats]{stats}}
}
