% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{confidence_angle}
\alias{confidence_angle}
\title{Confidence angle around the mean direction}
\usage{
confidence_angle(x, w = NULL, axial = TRUE, na.rm = TRUE, confidence = 0.95)
}
\arguments{
\item{x}{numeric vector. Values in degrees, for which the
mean, median or standard deviation are required.}

\item{w}{(optional) Weights. A vector of positive numbers, of the same length as
\code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(TRUE, the default) or circular, i.e. 2pi-periodical (FALSE).}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}

\item{confidence}{Level of confidence (as fraction). \code{0.95} is the default.}
}
\value{
Angle in degrees
}
\description{
Probabilistic limit on the location of the true or population mean direction,
assuming that the estimation errors are normally distributed.
}
\details{
The confidence angle gives the interval, i.e. plus and minus the confidence angle,
around the mean direction of a particular sample, that contains the true or population
mean direction under a given confidence.
}
\examples{
finland_stria <- c(23, 27, 53, 58, 64, 83, 85, 88, 93, 99, 100, 105, 113, 113, 114, 117, 121, 123, 125, 126, 126, 126, 127, 127, 128, 128, 129, 132, 132, 132, 134, 135, 137, 144, 145, 145, 146, 153, 155, 155, 155, 157, 163, 165, 171, 172, 179, 181, 186, 190, 212)
confidence_angle(finland_stria, axial = FALSE)

data(san_andreas)
data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, ep, "right")
circular_mean(sa.por$azi.PoR, w = 1/san_andreas$unc)
confidence_angle(sa.por$azi.PoR, w = 1/san_andreas$unc)
}
\references{
Davis (1986) Statistics and data analysis in geology. 2nd ed., John Wiley & Sons.
}
\seealso{
\code{\link[=mean_resultant_length]{mean_resultant_length()}}, \code{\link[=sd_error]{sd_error()}}
}
