% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{conjugate_Q4}
\alias{conjugate_Q4}
\title{Conjugation of a Quaternion}
\usage{
conjugate_Q4(q, normalize = FALSE)
}
\arguments{
\item{q}{object of class \code{"quaternion"}}

\item{normalize}{logical. Whether a quaternion normalization should be applied (TRUE) or not (FALSE, the default).}
}
\value{
object of class \code{"quaternion"}
}
\description{
Inverse rotation given by conjugated quaternion
}
\keyword{internal}
