% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{cunif}
\alias{cunif}
\alias{rcunif}
\alias{dcunif}
\alias{pcunif}
\alias{qcunif}
\title{The Circular Uniform Distribution}
\usage{
rcunif(n, axial = FALSE)

dcunif(theta, axial = FALSE, log = FALSE)

pcunif(theta, axial = FALSE, lower.tail = TRUE, log.p = FALSE)

qcunif(p, axial = FALSE, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{n}{integer. Number of observations in degrees}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{theta}{numeric. Angular value in degrees}

\item{log, log.p}{logical. If \code{TRUE}, probabilities p are given as log(p).}

\item{lower.tail}{logical. If \code{TRUE} (default), probabilities are \eqn{P(\Theta \le \theta)},
otherwise \eqn{P(\Theta > \theta)}.}

\item{p}{numeric. Vector of probabilities with values in \eqn{[0,1]}{[0,1]}.}
}
\value{
\code{dcunif} gives the density,
\code{pcunif} gives the probability of circular uniform distribution function,
\code{rcunif} generates random deviates (in degrees), and
\code{qcunif} provides quantiles (in degrees).
}
\description{
Density, probability distribution function, quantiles, and random generation
for the circular uniform distribution.
}
\examples{
set.seed(1)
x <- rcunif(5)

dcunif(x)
dcunif(x, axial = TRUE)

pcunif(x)
qcunif(c(.25, .5, .75))
}
\seealso{
\link{wcauchy} and \link{vonmises}
}
