% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{ddistance}
\alias{ddistance}
\title{Helper function for distance of coordinates to grid point}
\usage{
ddistance(lat_g, lon_g, lat, lon)
}
\arguments{
\item{lat_g, lon_g}{coordinate of grid point (degree)}

\item{lat, lon}{coordinates of points (degree)}
}
\description{
Returns the distance between a locations and a grid point in km
}
\examples{
\dontrun{
  ddistance(lat_g = 20, lon_g = 12, lat = c(50, 30), lon = c(40, 32))
}
}
