% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{dist_greatcircle}
\alias{dist_greatcircle}
\title{Distance between points}
\usage{
dist_greatcircle(
  lat1,
  lon1,
  lat2,
  lon2,
  r = earth_radius(),
  method = c("haversine", "haversine2", "vincenty")
)
}
\arguments{
\item{lat1, lon1}{coordinate of point(s) 1 (degrees).}

\item{lat2, lon2}{coordinates of point(s) 2 (degrees).}

\item{r}{radius of the sphere (default = 6371.0087714 km, i.e. the radius of
the Earth)}

\item{method}{Formula for calculating great circle distance:
\describe{
\item{"haversine"}{Haversine formula (the default)}
\item{"haversine2"}{Haversine formula optimized for 64-bit floating-point numbers}
\item{"vincenty"}{Vincenty formula for an ellipsoid with equal major and minor axes}
}}
}
\description{
Returns the great-circle distance between a location and all grid point in km
}
\examples{
dist_greatcircle(lat1 = 20, lon1 = 12, lat2 = c(50, 30), lon2 = c(40, 32))
dist_greatcircle(lat1 = 20, lon1 = 12, lat2 = c(50, 30), lon2 = c(40, 32),
method = "haversine2")
dist_greatcircle(lat1 = 20, lon1 = 12, lat2 = c(50, 30), lon2 = c(40, 32),
method = "vincenty")
}
\seealso{
\code{\link[=orthodrome_haversine]{orthodrome_haversine()}}, \code{\link[=orthodrome_haversine2]{orthodrome_haversine2()}}, \code{\link[=orthodrome_vincenty]{orthodrome_vincenty()}}
}
