% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{distance_from_pb}
\alias{distance_from_pb}
\title{Distance from plate boundary}
\usage{
distance_from_pb(x, ep, pb, tangential = FALSE)
}
\arguments{
\item{x, pb}{\code{sf} objects of the data points and the plate boundary
geometries in the geographical coordinate system}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole (\code{lat}, \code{lon})}

\item{tangential}{Logical. \code{TRUE} for tangential boundaries and
\code{FALSE} (the default) for inward and outward boundaries.}
}
\value{
Numeric vector of the (great circle) distances in degree
}
\description{
Distance of data points from the nearest plate boundary in degree
}
