% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_distance.R
\name{distance_mod}
\alias{distance_mod}
\title{Normalize angular distance on a sphere distance}
\usage{
distance_mod(x)
}
\arguments{
\item{x}{numeric, angular distance (in degrees)}
}
\value{
numeric vector
}
\description{
Helper function to express angular distance on the sphere in
the range of 0 to 180 °
}
\keyword{internal}
