% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{dummy}
\alias{dummy}
\alias{smallcircle_dummy}
\alias{greatcircle_dummy}
\alias{loxodrome_dummy}
\title{Plate Stress Dummy Grid}
\usage{
smallcircle_dummy(n)

greatcircle_dummy(n)

loxodrome_dummy(n, angle, cw)
}
\arguments{
\item{n}{Number of curves}

\item{angle}{Direction of loxodromes (in degree)}

\item{cw}{logical. Sense of loxodromes: \code{TRUE} for clockwise
loxodromes (right-lateral displaced plate boundaries). \code{FALSE} for
counterclockwise loxodromes (left-lateral displaced plate boundaries).}
}
\value{
\code{data.frame}
}
\description{
Helper functions to create a dummy grid for small circles,
great circles, and loxodromes of an Euler pole
}
\keyword{internal}
