% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{euler_from_rot}
\alias{euler_from_rot}
\title{Euler axis and angle from Euler matrix}
\usage{
euler_from_rot(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
\code{list} with the following objects
\describe{
\item{pole}{object of class \code{"euler.pole"}}
\item{psi}{numeric Euler angle in degree}
}
}
\description{
Extracts the coordinates of an Euler pole and the angle from a
Euler matrix
}
\details{
If there is no rotation (i.e. angle i zero), the coordinates of the
axis are equal to Earth's spin axis (according to the GPLATES convention).
}
\examples{
ep <- euler_pole(90, 0)
er <- euler_rot(ep, psi = 45)

euler_from_rot(er)
}
