% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{euler_rot}
\alias{euler_rot}
\title{Euler rotation matrix}
\usage{
euler_rot(ep, psi)
}
\arguments{
\item{ep}{An object of class \code{"euler.pole"}, or a data.frame
(containing lon and lat)}

\item{psi}{Angle of rotation in degree}
}
\value{
\code{matrix}
}
\description{
Creates a matrix from the given set of values.
}
\examples{
ep <- euler_pole(90, 0)
euler_rot(ep, psi = 45)
}
\references{
Greiner, B. (1999). Euler rotations in plate-tectonic
reconstructions. \emph{Computers and Geosciences}, 25(3), 209--216.
\doi{10.1016/S0098-3004(98)00160-5}
}
