% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{euler_to_Q4}
\alias{euler_to_Q4}
\title{Quaternion from Euler angle-axis representation for rotations}
\usage{
euler_to_Q4(x, normalize = FALSE)
}
\arguments{
\item{x}{\code{"euler.pole"} object}

\item{normalize}{logical. Whether a quaternion normalization should be applied (TRUE) or not (FALSE, the default).}
}
\value{
object of class \code{"quaternion"}
}
\description{
Quaternion from Euler angle-axis representation for rotations
}
