% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{eulerpole_loxodromes}
\alias{eulerpole_loxodromes}
\title{Draws loxodromes directed towards an Euler pole}
\usage{
eulerpole_loxodromes(x, angle = 45, ld = 10, sense, loxodrome = NULL)
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler pole in lat and lon}

\item{angle}{Direction of loxodromes; default = 45}

\item{ld}{numeric, angle between loxodromes (in degree); default: 10}

\item{sense}{Sense of loxodromes  'sinistral' or 'dextral' for 'clockwise' or 'counterclockwise' loxodromes, respectively}
}
\value{
An object of class \code{SpatialLinesDataFrame}
}
\description{
Construct curves cutting small circles at a constant angle
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to Pacific plate
eulerpole_loxodromes(x = euler, angle = 45, ld = 10, sense = "sinistral")
}
\author{
Tobias Stephan
}
