% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{geographical_to_cartesian}
\alias{geographical_to_cartesian}
\title{Geographical to cartesian coordinates}
\usage{
geographical_to_cartesian(p, r = 1)
}
\arguments{
\item{p}{vector of two numbers (latitude, longitude)}

\item{r}{radius. default is 1.}
}
\value{
Vector of three numbers (x, y, z)
}
\description{
Converts vector from geographical into cartesian  coordinates
}
\examples{
u <- c(50, 10)
geographical_to_cartesian(u)
}
