% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_distance.R
\name{get_distance}
\alias{get_distance}
\title{Helper function to Distance from plate boundary}
\usage{
get_distance(lon, lat, pb.coords, tangential, km)
}
\arguments{
\item{lon, lat}{numeric vectors}

\item{pb.coords}{matrix}

\item{tangential, km}{logical}
}
\description{
Helper function to Distance from plate boundary
}
\seealso{
\code{\link[=distance_from_pb]{distance_from_pb()}}
}
