% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{get_distance_mod}
\alias{get_distance_mod}
\title{Helper funtcion to get normalized distance from plate boundary}
\usage{
get_distance_mod(x)
}
\arguments{
\item{x}{numeric}
}
\description{
Helper funtcion to get normalized distance from plate boundary
}
\seealso{
\code{\link[=distance_from_pb]{distance_from_pb()}}
}
