% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_distance.R
\name{get_projected_pb_strike}
\alias{get_projected_pb_strike}
\title{Helper function to get Distance from plate boundary}
\usage{
get_projected_pb_strike(lon, lat, pb.coords, pb.bearing, tangential)
}
\arguments{
\item{lon, lat, pb.bearing}{numeric vectors}

\item{pb.coords}{matrix}

\item{tangential}{logical}
}
\description{
Helper function to get Distance from plate boundary
}
\seealso{
\code{\link[=projected_pb_strike]{projected_pb_strike()}}
}
\keyword{internal}
