% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{get_relrot}
\alias{get_relrot}
\title{Helper function to Equivalent rotation}
\usage{
get_relrot(plate.rot, lat, lon, angle, fixed, fixed.ep)
}
\arguments{
\item{plate.rot, fixed}{character or numeric}

\item{lat, lon, angle}{numeric}

\item{fixed.ep}{data.frame}
}
\description{
Helper function to Equivalent rotation
}
\seealso{
\code{\link[=equivalent_rotation]{equivalent_rotation()}}
}
