% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stress.R
\docType{data}
\name{homing}
\alias{homing}
\title{Homing directions of Pidgeons}
\format{
numeric vector of length 10
}
\description{
Directions of 10 pidgeons, example data from Mardia and Jupp (1999), pp. 93
}
\examples{
data("homing")
homing
}
\references{
Mardia and Jupp (1999). Directional Statistics. John Wiley and
Sons.
}
\keyword{datasets}
