% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stress.R
\docType{data}
\name{iceland}
\alias{iceland}
\title{Example stress data for outward-moving displaced plate boundary}
\format{
An object of class \code{"sf"}
}
\source{
\url{https://www.world-stress-map.org/}
}
\usage{
data('iceland')
}
\description{
Subset of the World Stress Map (WSM) compilation of information on the
crustal present-day stress field (Version 1.1. 2019). Subset contains stress data
of Iceland.
}
\examples{
data("iceland")
head(iceland)
}
\references{
Heidbach, O., M. Rajabi, X. Cui, K. Fuchs, B. M<U+00FC>ller, J.
Reinecker, K. Reiter, M. Tingay, F. Wenzel, F. Xie, M. O. Ziegler,
M.-L. Zoback, and M. D. Zoback (2018): The World Stress Map database
release 2016: Crustal stress pattern across scales. \emph{Tectonophysics},
\strong{744}, 484-498, \doi{10.1016/j.tecto.2018.07.007}.
}
\seealso{
\code{\link[=download_WSM2016]{download_WSM2016()}} for description of columns and stress regime
acronyms
}
\keyword{datasets}
