% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{is.euler}
\alias{is.euler}
\title{Check if object is euler.pole}
\usage{
is.euler(x)
}
\arguments{
\item{x}{Object of class \code{"data.frame"} containing the Euler poles of
plate rotations:
\describe{
\item{\code{plate.rot}}{Moving plate}
\item{\code{lat}, \code{lon}}{coordinates of Euler pole}
\item{\code{angle}}{Angle of rotation}
\item{\code{plate.fix}}{Fixed plate}
}}
}
\value{
logical
}
\description{
Check if object is euler.pole
}
