% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{is.Q4}
\alias{is.Q4}
\title{Check if object is quaternion}
\usage{
is.Q4(x)
}
\arguments{
\item{x}{object of class \code{"quaternion"}}
}
\value{
logical
}
\description{
Check if object is quaternion
}
\keyword{internal}
