% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{longitude_modulo}
\alias{longitude_modulo}
\title{Longitude Correction}
\usage{
longitude_modulo(longitude)
}
\arguments{
\item{longitude}{Longitude(s) in degree}
}
\value{
Number
}
\description{
Corrects the longitude value to values between -180 and +180
degree
}
\examples{
longitude_modulo(-361)
}
