% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{loxodrome_dummy}
\alias{loxodrome_dummy}
\title{Loxodrome dummy}
\usage{
loxodrome_dummy(angle, n, sense)
}
\arguments{
\item{angle}{Direction of loxodromes (in degree)}

\item{n}{number of loxodromes  (in  degree)}

\item{sense}{Sense of loxodromes 'sinistral' or 'dextral' for 'clockwise' or
'counterclockwise' loxodromes, respectively}
}
\value{
data.frame
}
\description{
create a dummy for loxodromes
}
\examples{
loxodrome_dummy(angle = 45, n = 10, sense = "sinistral")
}
