% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mean_SC}
\alias{mean_SC}
\title{Mean Cosine and Sine}
\usage{
mean_SC(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{numeric. Angles in degrees}

\item{w}{numeric. Weightings}

\item{na.rm}{logical. Remove \code{NA} values before calculation?}
}
\value{
named two element vector
}
\description{
Mean Cosine and Sine
}
\examples{
\dontrun{
set.seed(1)
x <- rvm(100, 0, 5)
mean_SC(x)
}
}
\keyword{internal}
