% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misfit_shmax.R
\name{misfit_shmax}
\alias{misfit_shmax}
\title{Deviation of observed and predicted directions of SHmax}
\usage{
misfit_shmax(prd, obs)
}
\arguments{
\item{prd}{data.frame containing the modeled azimuths of SHmax, i.e. the
return object from model_shmax()}

\item{obs}{numeric vector containing the observed azimuth of SHmax,
same length as prd}
}
\value{
data.frame containing the deviation angles between the  observed and
the modeled maximum horizontal stresses along great circles (dev.gc),
small circles (dev.sc), and loxodromes (dev.ld.cw, dev.ld.ccw)
}
\description{
Calculate the angular difference between the observed and modeled direction
of maximum horizontal stresses along great circles, small circles, and
loxodromes of the relative plate motion´s Euler pole
}
\author{
Copyright (C) 2021 Tobias Stephan
}
