% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{norm_rayleigh}
\alias{norm_rayleigh}
\title{Normalized goodness-of-fit test}
\usage{
norm_rayleigh(x, prd = NULL, unc, axial = TRUE)
}
\arguments{
\item{x}{numeric vector. Values in degrees}

\item{prd}{The a priori expected direction (in degrees) for the alternative
hypothesis.}

\item{unc}{numeric. The standard deviations of \code{x}. If \code{NULL}, the non-weighted
Rayleigh test is performed.}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}
}
\value{
a list with the components:
\describe{
\item{\code{statistic}}{Test statistic}
\item{\code{p.value}}{significance level of the test statistic}
}
}
\description{
Weighted version of the Rayleigh test (or V-test) for uniformity around a
a priori expected von Mises concentration
}
\details{
The Null hypothesis is that the distributions of the observations \code{x} is an
unimodal distribution with a specified mean direction (\code{prd}).
If \code{statistic > p.value}, the null hypothesis is rejected.
If not, the uniform distribution cannot be excluded.
}
\examples{
# Load data
data("cpm_models")
data(san_andreas)
ep <- equivalent_rotation(subset(cpm_models, model == "NNR-MORVEL56"), "na", "pa")
sa.por <- PoR_shmax(san_andreas, ep, "right")
data("iceland")
ep.ice <- equivalent_rotation(subset(cpm_models, model == "NNR-MORVEL56"), "eu", "na")
ice.por <- PoR_shmax(iceland, ep.ice, "out")
data("tibet")
ep.tib <- equivalent_rotation(subset(cpm_models, model == "NNR-MORVEL56"), "eu", "in")
tibet.por <- PoR_shmax(tibet, ep.tib, "in")

# GOF test:
norm_rayleigh(tibet.por$azi.PoR, prd = 90, unc = tibet$unc)
norm_rayleigh(ice.por$azi.PoR, prd = 0, unc = iceland$unc)
norm_rayleigh(sa.por$azi.PoR, prd = 135, unc = san_andreas$unc)
}
\seealso{
\code{\link[=rayleigh_test]{rayleigh_test()}}
}
