% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{norm_rayleigh_test}
\alias{norm_rayleigh_test}
\title{Normalized Rayleigh goodness-of-fit test}
\usage{
norm_rayleigh_test(x, unc, prd = NULL, axial = TRUE)
}
\arguments{
\item{x}{numeric vector. Values in degrees}

\item{unc}{numeric. The standard deviations of \code{x}. If \code{NULL}, the non-weighted
Rayleigh test is performed.}

\item{prd}{The specified or known mean direction (in degrees) in alternative hypothesis}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}
}
\value{
a list with the components:
\describe{
\item{\code{statistic}}{Test statistic}
\item{\code{p.value}}{significance level of the test statistic}
}
}
\description{
Performs a Rayleigh test of uniformity using observations weighted by
their uncertainties.
}
\details{
The Null hypothesis is that the distributions of the observations \code{x} is an
unimodal distribution with a specified mean direction (\code{prd}).
If \code{statistic > p.value}, the null hypothesis is rejected.
If not, randomness (uniform distribution) cannot be excluded.
}
\examples{
data("cpm_models")
data(san_andreas)
ep <- equivalent_rotation(cpm_models \%>\% filter(model == "NNR-MORVEL56"), "na", "pa")
sa.por <- PoR_shmax(san_andreas, ep, "right")
data("iceland")
ep.ice <- equivalent_rotation(cpm_models \%>\% filter(model == "NNR-MORVEL56"), "eu", "na")
ice.por <- PoR_shmax(iceland, ep.ice, "out")
data("tibet")
ep.tib <- equivalent_rotation(cpm_models \%>\% filter(model == "NNR-MORVEL56"), "eu", "in")
tibet.por <- PoR_shmax(tibet, ep.tib, "in")

norm_rayleigh_test(tibet.por$azi.PoR, unc = tibet$unc, prd = 90)
norm_rayleigh_test(ice.por$azi.PoR, unc = iceland$unc, prd = 0)
norm_rayleigh_test(sa.por$azi.PoR, unc = san_andreas$unc, prd = 135)
}
\seealso{
\code{\link[=rayleigh_test]{rayleigh_test()}}
}
