% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{normalize_Q4}
\alias{normalize_Q4}
\title{Quaternion normalization}
\usage{
normalize_Q4(q)
}
\arguments{
\item{q}{quaternion}
}
\value{
object of class \code{"quaternion"}
}
\description{
Quaternion normalization
}
