% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuvel1_plates.R
\docType{data}
\name{nuvel1_plates}
\alias{nuvel1_plates}
\title{Plate Boundaries on the Earth}
\format{
An object of class \code{"sf"}
}
\usage{
data('nuvel1_plates')
}
\description{
Global set of present plate boundaries on the Earth based on NUVEL-1 model by
DeMets et al. 1990
}
\examples{
data("nuvel1_plates")
head("nuvel1_plates")
}
\references{
DeMets, C., Gordon, R. G., Argus, D. F., & Stein, S. (1990).
Current plate motions. \emph{Geophysical Journal International}, 101(2), 425--478.
\doi{10.1111/j.1365-246X.1990.tb06579.x}.
}
\keyword{datasets}
