% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ot_eigen2d}
\alias{ot_eigen2d}
\title{Decompositon of orientation tensor}
\usage{
ot_eigen2d(x, w = NULL, axial = TRUE)
}
\arguments{
\item{x}{numeric vector. Values in degrees.}

\item{w}{(optional) Weights. A vector of positive numbers and of the same
length as \code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}
}
\description{
Decompositon of orientation tensor
}
\examples{
test <- rvm(100, mean = 0, k = 19)
ot_eigen2d(test, axial = FALSE)

data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
sa_eig <- ot_eigen2d(sa.por$azi.PoR, w = weighting(san_andreas$unc))
print(sa_eig)

rose(sa.por$azi.PoR)
rose_line(sa_eig$vectors, col = c('blue', 'green'))
}
