% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB2002.R
\docType{data}
\name{PB2002}
\alias{PB2002}
\title{Global set of present plate boundaries on the Earth (PB2002)}
\format{
An object of class \code{""SpatialPolygonsDataFrame""}
}
\usage{
data(PB2002)
}
\description{
Global set of present plate boundaries on the Earth (PB2002) by Bird 2003
}
\examples{

data(PB2002)
head(PB2002)
}
\references{
Bird, P. (2003), An updated digital model of plate boundaries,
Geochem. Geophys. Geosyst., 4, 1027, doi:10.1029/2001GC000252, 3.
}
\keyword{datasets}
