% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb2002_plates.R
\docType{data}
\name{pb2002_plates}
\alias{pb2002_plates}
\title{Plate Boundaries on the Earth}
\format{
An object of class \code{"sf"}
}
\usage{
data('pb2002_plates')
}
\description{
Global set of present plate boundaries on the Earth (PB2002) by Bird 2003
}
\examples{
data("pb2002_plates")
head("pb2002_plates")
}
\references{
Bird, P. (2003), An updated digital model of plate boundaries,
\emph{Geochem. Geophys. Geosyst.}, 4, 1027,\doi{10.1029/2001GC000252}, 3.
}
\keyword{datasets}
