% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\docType{data}
\name{plates}
\alias{plates}
\title{Plate Boundaries on the Earth}
\format{
An object of class \code{"sf"}
}
\usage{
data('plates')
}
\description{
Global set of present plate boundaries on the Earth based on PB2002 model by
Bird (2003).
Contains the plate boundary displacement types such as inward, outward, or
tangentially displacement.
}
\examples{
data("plates")
head("plates")
}
\references{
Bird, P. (2003), An updated digital model of plate boundaries,
\emph{Geochem. Geophys. Geosyst.}, 4, 1027, doi: 10.1029/2001GC000252, 3.
}
\keyword{datasets}
