% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateTectonicStressR.R
\docType{package}
\name{PlateTectonicStressR}
\alias{PlateTectonicStressR}
\alias{_PACKAGE}
\alias{PlateTectonicStressR-package}
\title{library(PlateTectonicStressR)}
\description{
A list of documented functions may be viewed by typing
\code{help(package='PlateTectonicStressR')}. Further
details about the theoretical background are provided by
Wdowinski (1999).
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, http://dx.doi.org/10.1029/97JB03390.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/PlateTectonicStressR/}
  \item Report bugs at \url{https://github.com/tobiste/PlateTectonicStressR/issues}
}

}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

}
