% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR2Geo_shmax}
\alias{PoR2Geo_shmax}
\title{Azimuth conversion from PoR to geographical coordinate reference system}
\usage{
PoR2Geo_shmax(x, PoR)
}
\arguments{
\item{x}{\code{data.frame} containing the PoR equivalent azimuths
(\code{azi.PoR}), and either the geographical coordinates of the
point(s) or the PoR-equivalent coordinates.}

\item{PoR}{\code{data.frame} containing the geographical location of
the Euler pole (\code{lat}, \code{lon})}
}
\description{
Helper function to convert PoR azimuths into geographical azimuths
}
\examples{
data("nuvel1")
# North America relative to Pacific plate:
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
data("san_andreas")
head(san_andreas$azi)
san_andreas$azi.PoR <- PoR_shmax(san_andreas, PoR)
res.geo <- PoR2Geo_shmax(san_andreas, PoR)
head(res.geo)
}
