% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR_azimuth}
\alias{PoR_azimuth}
\title{Azimuth relative to PoR}
\usage{
PoR_azimuth(x, ep)
}
\arguments{
\item{x}{\code{"data.frame"}. coordinates of data point (lat, lon) and the azimuth}

\item{ep}{\code{"data.frame"}. coordinates of Euler pole (PoR)}
}
\value{
\code{"numeric"}. Transformed azimuth relative to PoR
}
\description{
Transform azimuth from geographical to PoR system according to Wdowinski 1998
}
\examples{
data("nuvel1")
# North America relative to Pacific plate:
euler <- subset(nuvel1, nuvel1$plate.rot == "na")

data("san_andreas")
res <- PoR_azimuth(san_andreas, euler)
head(res)
}
