% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{por_conversion}
\alias{por_conversion}
\alias{PoR_to_geographical2}
\alias{geographical_to_PoR2}
\title{Conversion between PoR to geographical coordinate system}
\usage{
PoR_to_geographical2(x, ep)

geographical_to_PoR2(x, ep)
}
\arguments{
\item{x}{two-colum vector of the Point in the PoR system (colatitude, azimuth) or the geographical coordinate system (latitude, longitude)}

\item{ep}{two-colum Numeric vector containing the geographic location of the Euler
pole (latitude, longitude)}
}
\value{
two-colum vector of the transformed latitude, longtitude or colatitude, azimuth in geographical or PoR coordinate system, respectively
}
\description{
Transformation from PoR to geographical coordinate system or vice versa
}
\examples{
\dontrun{
ep.geo <- c(52, 50)
q.geo <- c(10, 45)
q.por <- geographical_to_PoR2(q.por, ep.geo)
PoR_to_geographical2(q.por, ep.geo)
}
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, \doi{10.1029/97JB03390}.
}
