% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_conversion_df}
\alias{por_conversion_df}
\alias{PoR_to_geographical2}
\alias{geographical_to_PoR2}
\title{Conversion between spherical PoR to geographical coordinate system}
\usage{
PoR_to_geographical2(x, euler, spherical = TRUE)

geographical_to_PoR2(x, euler, spherical = TRUE)
}
\arguments{
\item{x}{\code{sf} or \code{data.frame} containing (co)lat and lon coordinates
(\code{lat}, \code{lon}) of the points to be transformed}

\item{euler}{\code{data.frame} of the geographical coordinates of the Euler pole
(\code{(co)lat}, \code{lon})}

\item{spherical}{logical. Whether x or the return are in spherical
coordinates, i.e. colat-lon (TRUE, the default), or in lat-lon (FALSE)}
}
\value{
\code{data.frame} with the PoR coordinates
(\code{colat.PoR}, \code{lon.PoR} or \code{lat.PoR}, \code{lon.PoR})
}
\description{
Transformation from spherical PoR to geographical coordinate system and
vice versa
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- geographical_to_PoR2(san_andreas, euler)
head(san_andreas.por)
head(PoR_to_geographical2(san_andreas.por, euler))
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, \doi{10.1029/97JB03390}.
}
