% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_conversion_vec}
\alias{por_conversion_vec}
\alias{PoR_to_geographical_vec}
\alias{geographical_to_PoR_vec}
\title{Conversion between PoR to geographical coordinate system}
\usage{
PoR_to_geographical_vec(x, euler, spherical = TRUE)

geographical_to_PoR_vec(x, euler, spherical = TRUE)
}
\arguments{
\item{x, euler}{two-column vectors containing the (co)lat and lon coordinates}

\item{spherical}{logical. Whether x or the return are in spherical
coordinates}
}
\description{
Helper function for the transformation from PoR to geographical coordinate
system or vice versa
}
\examples{
\dontrun{
ep.geo <- c(20, 33)
q.geo <- c(10, 45)
q.por <- geographical_to_PoR_vec(q.geo, ep.geo)
q.por
PoR_to_geographical_vec(q.por, ep.geo)
}
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, \doi{10.1029/97JB03390}.
}
