% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{PoR_coordinates}
\alias{PoR_coordinates}
\title{Coordinates of the Pole of Rotation Reference System}
\usage{
PoR_coordinates(x, PoR)
}
\arguments{
\item{x}{Can be either a \code{"data.frame"} containing \code{lat} and \code{lon}
coordinates of a point in the geographical CRS or the \code{lat.PoR},
\code{lon.PoR}) of the point in the PoR CRS,
a two-column matrix containing the lat and lon coordinates,
a \code{sf} object, or a \code{raster} object.}

\item{PoR}{Pole of Rotation. \code{"data.frame"} containing the geographical coordinates of the Euler pole}
}
\value{
\code{\link[=PoR_coordinates]{PoR_coordinates()}} returns \code{data.frame} with the PoR coordinates
(\code{lat.PoR}, \code{lon.PoR}).
}
\description{
Retrieve the PoR equivalent coordinates of an object
}
\examples{
data("nuvel1")
por <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")

# coordinates from sf object
san_andreas.por_sf <- PoR_coordinates(san_andreas, por)
head(san_andreas.por_sf)

# coordinates from data.frame
san_andreas.por_df <- PoR_coordinates(sf::st_drop_geometry(san_andreas), por)
head(san_andreas.por_df)
}
