% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_crs}
\alias{por_crs}
\alias{PoR_to_geographical}
\alias{geographical_to_PoR}
\title{Conversion between PoR to geographical coordinate system}
\usage{
PoR_to_geographical(x, ep)

geographical_to_PoR(x, ep)
}
\arguments{
\item{x}{\code{sf} object of the data points in geographical or PoR coordinate system}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole (\code{lat}, \code{lon})}
}
\value{
\code{sf} object of the data points in the transformed geographical or PoR
coordinate system
}
\description{
Transform spherical objects from PoR to geographical coordinate system and
vice versa.
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate

data("wsm2016")
example.geo <- sf::st_set_crs(sf::st_as_sf(wsm2016[1:10, ], coords = c("lon", "lat")), "EPSG:4326")

example.por <- geographical_to_PoR(example.geo, euler)
PoR_to_geographical(example.por, euler)
}
