% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{PoR_distance}
\alias{PoR_distance}
\title{Distance to Pole of Rotation}
\usage{
PoR_distance(x, PoR, FUN = orthodrome)
}
\arguments{
\item{x}{Can be either a \code{"data.frame"} containing \code{lat} and \code{lon}
coordinates of a point in the geographical CRS or the \code{lat.PoR},
\code{lon.PoR}) of the point in the PoR CRS,
a two-column matrix containing the lat and lon coordinates,
a \code{sf} object, or a \code{raster} object.}

\item{PoR}{Pole of Rotation. \code{"data.frame"} containing the geographical coordinates of the Euler pole}

\item{FUN}{function to calculate the great-circle distance.
\code{\link[=orthodrome]{orthodrome()}}, \code{\link[=haversine]{haversine()}} (the default), or \code{\link[=vincenty]{vincenty()}}.}
}
\value{
numeric vector. Great-circle distance in degree
}
\description{
Retrieve the (angular) great-circle distance between a point and the PoR (Euler pole).
}
\examples{
data("nuvel1")
por <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")

# distance form sf object
PoR_distance(san_andreas, por) |> head()

# distance form data.frame
PoR_distance(sf::st_drop_geometry(san_andreas), por) |> head()
PoR_distance(sf::st_drop_geometry(san_andreas), por, FUN = orthodrome) |> head()
PoR_distance(sf::st_drop_geometry(san_andreas), por, FUN = vincenty) |> head()
}
