% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PoR_plot}
\alias{PoR_plot}
\title{Plotting the \eqn{\sigma_{Hmax}}{SHmax} azimuth}
\usage{
PoR_plot(azi, distance, prd, unc = NULL, regime, width = 51, ...)
}
\arguments{
\item{azi}{numeric. Azimuth of \eqn{\sigma_{Hmax}}{SHmax}}

\item{distance}{numeric. Distance to plate boundary}

\item{prd}{numeric. the predicted direction of \eqn{\sigma_{Hmax}}{SHmax}}

\item{unc}{numeric. Uncertainty of observed \eqn{\sigma_{Hmax}}{SHmax}, either a
numeric vector or a number}

\item{regime}{character vector. The stress
regime (following the classification of the World Stress Map)}

\item{width}{integer. window width (in number of observations) for moving average
of the azimuths and Norm Chi-square statistics.}

\item{...}{optional arguments to \code{zoo::rollapply()}, \code{\link[=rose]{rose()}}, or \code{plot()}}
}
\description{
Creates a set of plots including
the azimuth as a function of the distance to the plate boundary,
the Norm Chi-squared as a function of the distance to the plate boundary,
and a rose diagram of the frequency distribution of the azimuths.
}
\examples{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")

data("san_andreas")
res <- PoR_shmax(san_andreas, na_pa, "right")
d <- distance_from_pb(san_andreas, na_pa, plate_boundary, tangential = TRUE)
PoR_plot(res$azi.PoR, d, res$prd, san_andreas$unc, san_andreas$regime)
}
\seealso{
\code{\link[=PoR_shmax]{PoR_shmax()}}, \code{\link[=distance_from_pb]{distance_from_pb()}}, \code{\link[=circular_median]{circular_median()}}, \code{\link[=circular_IQR]{circular_IQR()}}, \code{\link[=norm_chisq]{norm_chisq()}}
}
