% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR_shmax}
\alias{PoR_shmax}
\title{Theoretical Direction of Maximum Horizontal Stress in PoR reference system}
\usage{
PoR_shmax(df, euler, type = c("in", "out", "right", "left"))
}
\arguments{
\item{df}{\code{data.frame} containing the coordinates of the point(s)
(\code{lat}, \code{lon}), the orientation of
\eqn{\sigma_{\text{Hmax}}}{SHmax} \code{azi} and its standard deviation
\code{unc} (optional)}

\item{euler}{\code{data.frame} containing the coordinates of the Euler pole
for the plate boundary  (\code{lat}, \code{lon})}

\item{type}{Character. Type of plate boundary (optional). Can be
\code{"out"}, \code{"in"}, \code{"right"}, or
\code{"left"} for outward, inward, right-lateral, or left-lateral
moving plate boundaries, respectively.}
}
\value{
Either a numeric vector of the azimuths in the transformed coordinate
system, or a \code{"data.frame"} with
\describe{
\item{\code{"azi.PoR"}}{the transformed azimuths,}
\item{\code{"prd"}}{the predicted azimuths,}
\item{\code{"dev"}}{the deviation, and}
\item{\code{"nchisq"}}{the normalized \eqn{\chi^2}{chi-squared} statistics.}
}
}
\description{
Models the direction of maximum horizontal stress
\eqn{\sigma_{\text{Hmax}}}{SHmax} in the Euler pole (Pole of Rotation)
cooridnate reference system. When type of plate boundary is given, it also
gives the deviation from the theoretically predicted azimuth of
\eqn{\sigma_{\text{Hmax}}}{SHmax}, the deviation, and the normalized
\eqn{\chi^2}{chi-squared} statistics.
}
\details{
According to the theory, the azimuth of
\eqn{\sigma_{\text{Hmax}}}{SHmax} in the pole of rotation reference system is
approximate 0 (or 180), 45, 90, 135 degrees if the stress is sourced by an
outward, sinistral, inward, or dextral moving plate boundary, respectively.
directions of \eqn{\sigma_{\text{Hmax}}}{SHmax} with respect to the four
plate boundary types.
}
\examples{
data("nuvel1")
# North America relative to Pacific plate:
euler <- subset(nuvel1, nuvel1$plate.rot == "na")

data("wsm2016")
stress <- subset(
  wsm2016,
  wsm2016$lat >= 23 & wsm2016$lat <= 40 &
    wsm2016$lon >= -126 & wsm2016$lon <= -108
)
stress$unc <- stress$sd
res <- PoR_shmax(stress, euler, type = "right")
head(res)
}
\seealso{
\code{\link[=model_shmax]{model_shmax()}} to compute the theoretical orientation of
\eqn{\sigma_\text{Hmax}}{SHmax} in the geographical reference system.
\code{\link[=misfit_shmax]{misfit_shmax()}} to compute the deviation of the modeled direction
from the observed direction of \eqn{\sigma_\text{Hmax}}{SHmax}.
\code{\link[=norm_chisq]{norm_chisq()}} to calculate the normalized \eqn{\chi^2}{chi-squared}
statistics.
}
