% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{product_Q4}
\alias{product_Q4}
\title{Product of quaternions}
\usage{
product_Q4(q1, q2, normalize = FALSE)
}
\arguments{
\item{q1, q2}{two objects of class \code{"quaternion"}. first rotation R1 expressed by q1 followed by second rotation R2 expressed by q2}

\item{normalize}{logical. Whether a quaternion normalization should be applied (TRUE) or not (FALSE, the default).}
}
\value{
object of class \code{"quaternion"}
}
\description{
Helper function for multiplication of two quaternions.
Concatenation of two rotations R1 followed by R2
}
\note{
Multiplication is not commutative.
}
\keyword{internal}
