% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{Q4_to_euler}
\alias{Q4_to_euler}
\title{Euler angle/axis from quaternion}
\usage{
Q4_to_euler(q)
}
\arguments{
\item{q}{object of class \code{"quaternion"}}
}
\value{
\code{"euler.pole"} object
}
\description{
Euler angle/axis from quaternion
}
\keyword{internal}
