% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{quantise_wsm_quality}
\alias{quantise_wsm_quality}
\title{Quantize World Stress Map quality ranking}
\usage{
quantise_wsm_quality(x)
}
\arguments{
\item{x}{Either a string or a character vector of WSM quality ranking}
}
\value{
\code{"integer"} or vector of type \code{"integer"}
}
\description{
Quantize the World Stress Map A, B, C, D quality ranking
}
\examples{
quantise_wsm_quality(c("A", "B", "C", "D", NA))
data("wsm2016")
quantise_wsm_quality(wsm2016$quality)
}
\references{
Heidbach, O.; Barth, A.; Müller, B.; Reinecker, J.;
Stephansson, O.; Tingay, M.; Zang, A. (2016). WSM quality
ranking scheme, database description and analysis guidelines for stress
indicator. World Stress Map Technical Report 16-01, GFZ German Research
Centre for Geosciences. DOI: http://doi.org/10.2312/wsm.2016.001
}
