% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{quick_analysis}
\alias{quick_analysis}
\title{Quick analysis of a stress data set}
\usage{
quick_analysis(
  x,
  euler,
  type = c("none", "in", "out", "right", "left"),
  pb,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} or \code{sf} object containing the coordinates of the point(s)
(\code{lat}, \code{lon}), the orientation of
\eqn{\sigma_{Hmax}}{SHmax} \code{azi} and its standard deviation
\code{unc} (optional)}

\item{euler}{\code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler pole}

\item{type}{Character. Type of plate boundary (optional). Can be
\code{"out"}, \code{"in"}, \code{"right"}, or
\code{"left"} for outward, inward, right-lateral, or left-lateral
moving plate boundaries, respectively. If \code{"none"} (the default), only
the PoR-equivalent azimuth is returned.}

\item{pb}{(optional) \code{sf} object of the plate boundary geometries in the geographical
coordinate system}

\item{plot}{(logical). Whether to produce a plot additional to output.}

\item{...}{optional arguments to \code{\link[=distance_from_pb]{distance_from_pb()}}}
}
\value{
list. results of the coordinate and azimuth conversion, deviation,
misfit to predicted stress orientation and, if given, distance to tested
plate boundary as well as the normalized Chi-squared test statistic.
}
\description{
Returns the converted azimuths, distances to the plate boundary,
statistics of the model, and some plots.
}
\examples{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")

data("san_andreas")

quick_analysis(san_andreas, na_pa, type = "right", plate_boundary, plot = FALSE)
}
\seealso{
\code{\link[=PoR_shmax]{PoR_shmax()}}, \code{\link[=distance_from_pb]{distance_from_pb()}}, \code{\link[=norm_chisq]{norm_chisq()}}, \code{\link[=PoR_plot]{PoR_plot()}}
}
