% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{rose_bw}
\alias{rose_bw}
\alias{rose_bins}
\alias{rose_binwidth}
\title{Selecting optimal number of bins and width for rose diagrams}
\usage{
rose_bins(n, round = FALSE)

rose_binwidth(n, axial = TRUE, ...)
}
\arguments{
\item{n}{Integer. number of data}

\item{round}{Logical. Whether bin width is round to zero digits
(\code{round=TRUE}, the default)
or as is (\code{FALSE}).}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})
or biaxial (\code{TRUE}, the default).}

\item{...}{Additional arguments passed to \code{\link[=rose_bw]{rose_bw()}}.}
}
\description{
Selecting optimal number of bins and width for rose diagrams
}
\keyword{internal}
