% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rose_freq}
\alias{rose_freq}
\title{Helper Function to count frequencies and densities for rose diagrams}
\usage{
rose_freq(
  x,
  bins = NULL,
  ...,
  weights = NULL,
  binwidth = NULL,
  round_binwidth = 0,
  equal_area = TRUE,
  main = NULL,
  axial = TRUE
)
}
\arguments{
\item{x}{Data to be plotted. A numeric vector containing angles (in degrees).}

\item{bins}{number of arcs to partition the circle width.
Overridden by \code{binwidth}.}

\item{...}{Additional arguments passed to \code{\link[spatstat.explore:rose]{spatstat.explore::rose()}}.}

\item{weights}{Optional vector of numeric weights associated with x.}

\item{binwidth}{The width of the bins (in degrees).}

\item{round_binwidth}{integer. Number of decimal places of bin width (0 by
default).}

\item{equal_area}{Logical. Whether the radii of the bins are proportional to
the frequencies (\code{equal_area=FALSE}, i.e. equal-angle) or proportional to the
square-root of the frequencies (\code{equal_area=TRUE}, the default).}

\item{main}{description}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})
or biaxial (\code{TRUE}, the default).}

\item{breaks}{description}

\item{nclass}{description}

\item{start}{description}
}
\value{
an object of class \code{"histogram"}
}
\description{
Helper Function to count frequencies and densities for rose diagrams
}
\keyword{internal}
