% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{rotate_lines}
\alias{rotate_lines}
\title{Rotate Lines}
\usage{
rotate_lines(theta, p, centre)
}
\arguments{
\item{theta}{Angle of rotation (in degree)}

\item{p}{Points of lines}

\item{centre}{Center point of rotation}
}
\value{
\code{matrix}
}
\description{
Rotate a set of lines around a angle
}
