% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_angle}
\alias{rotation_angle}
\title{Rotation angle from rotation matrix}
\usage{
rotation_angle(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
numeric angle in degree
}
\description{
Extracts the rotation angle from rotation matrix
}
\note{
Infinitesimal small rotation (i.e. small angles) will cause an Error
due to round-off errors. In order to avoid the error, these rotations will
be treated as equal rotations. The function will print a warning message when
this is the case.
}
\examples{
w <- c(0, 1, 0)
rot <- rotation_matrix(w, 90)

rotation_angle(rot)
}
\seealso{
\code{\link[=rotation_axis]{rotation_axis()}} to extract the axis of the rotation matrix.
}
