% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_axis}
\alias{rotation_axis}
\title{Rotation axis from rotation matrix}
\usage{
rotation_axis(A)
}
\arguments{
\item{A}{3x3 matrix}
}
\value{
vector
}
\description{
Extracts the rotation axis from rotation matrix
}
\examples{
w <- c(0, 1, 0)
rot <- rotation_matrix(w, 90)

rotation_axis(rot)
}
\seealso{
\code{\link[=rotation_angle]{rotation_angle()}} to extract the angle of the rotation matrix.
}
