% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation matrix}
\usage{
rotation_matrix(n, alpha)
}
\arguments{
\item{n}{Rotation axis (in Cartesian coordinates). Can be a vector of three
numbers
or a matrix of 3 columns (first one is x, second y, third z)}

\item{alpha}{Rotation angle in degrees}
}
\value{
matrix
}
\description{
Calculates the rotation matrix using the rotation axis and the
angle of rotation
}
\details{
If \eqn{\vec u} is a vector prior to rotation and \eqn{\vec u'} is
the point after rotation then \eqn{\vec u' = R \cdot \vec u} where \eqn{R} is
a 3x3 rotation matrix:
\eqn{
  R = \begin{bmatrix} \cos \theta +u_{x}^{2}\left(1-\cos \theta \right) & u_{x}u_{y}\left(1-\cos \theta \right)-u_{z}\sin \theta & u_{x}u_{z}\left(1-\cos \theta \right)+u_{y}\sin \theta \\ u_{y}u_{x}\left(1-\cos \theta \right)+u_{z}\sin \theta & \cos \theta +u_{y}^{2}\left(1-\cos \theta \right)& u_{y}u_{z}\left(1-\cos \theta \right)-u_{x}\sin \theta \\ u_{z}u_{x}\left(1-\cos \theta \right)-u_{y}\sin \theta & u_{z}u_{y}\left(1-\cos \theta \right)+u_{x}\sin \theta & \cos \theta +u_{z}^{2}\left(1-\cos \theta \right)\end{bmatrix}
}
}
\examples{
w <- c(0, 1, 0)
rotation_matrix(w, 90)
}
\author{
Tobias Stephan
}
