% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_Q4}
\alias{rotation_Q4}
\title{Rotation of a vector by a quaternion}
\usage{
rotation_Q4(q, p)
}
\arguments{
\item{q}{object of class \code{"quaternion"}}

\item{p}{three-column vector (Cartesian coordinates) of unit length}
}
\value{
three-column vector (Cartesian coordinates) of unit length
}
\description{
Rotation of a vector by a quaternion
}
