% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_andreas.R
\docType{data}
\name{san_andreas}
\alias{san_andreas}
\title{Example stress data}
\format{
An object of class \code{"sf"}
}
\source{
\url{https://www.world-stress-map.org/}
}
\usage{
data('san_andreas')
}
\description{
Subset of the World Stress Map (WSM) compilation of information on the
crustal present-day stress field (Version 2016). Subset contains stress data
adjacent to the San Andreas Fault.
}
\examples{
data("san_andreas")
head("san_andreas")
}
\references{
Heidbach, O., M. Rajabi, X. Cui, K. Fuchs, B. Müller, J.
Reinecker, K. Reiter, M. Tingay, F. Wenzel, F. Xie, M. O. Ziegler,
M.-L. Zoback, and M. D. Zoback (2018): The World Stress Map database
release 2016: Crustal stress pattern across scales. \emph{Tectonophysics}, 744,
484--498, \doi{10.1016/j.tecto.2018.07.007}.
}
\keyword{datasets}
