% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{stresstrain}
\alias{stresstrain}
\alias{displacement_vector}
\alias{stress_matrix}
\title{Displacement vector and stress matrix in PoR}
\usage{
displacement_vector(x, ep, tangential = FALSE, positive = TRUE)

stress_matrix(x, ep, tangential = FALSE, positive = FALSE, v = 0.25, E = 50)
}
\arguments{
\item{x}{\code{sf} object of the data points in the geographical coordinate system}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole
(\code{lat}, \code{lon})}

\item{tangential}{Logical Whether the plate boundary is a tangential
boundary (\code{TRUE}) or an inward and outward boundary (\code{FALSE}, the
default).}

\item{positive}{Sign of the equatorial relative plate boundary motion (displacement).
If \code{tangential == TRUE}, positive values indicate right-lateral plate boundary displacement and negative values represent left-lateral displacement.
If \code{tangential == FALSE}, positive values indicate outward-moving plate boundary offset and negative values represent inward-moving plate boundary displacement.}

\item{v}{Numeric. Poisson's ratio. default is 0.25}

\item{E}{Numeric. Young's modulus. default is 50 (GPa)}
}
\value{
\code{sf} object of the data points in the PoR coordinate system
and the components of the displacement vector or the stress matrix.
}
\description{
Displacement vector and stress matrix in PoR
}
\examples{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")

data("san_andreas")
res <- displacement_vector(x = san_andreas, ep = na_pa, tangential = TRUE, positive = FALSE)
head(res)

res2 <- stress_matrix(x = san_andreas, ep = na_pa, tangential = TRUE, positive = FALSE)
head(res2)
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics. \emph{Journal of Geophysical Research: Solid Earth}, v. 103, p.
5037-5059, doi: 10.1029/97JB03390.
}
