% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stress.R
\docType{data}
\name{striae}
\alias{striae}
\title{Glacial striae from Finland}
\format{
numeric vector of length 51
}
\description{
Directions of 51 glacial striations, example data from Davis (1986), pp. 316
}
\examples{
data("striae")
striae
}
\references{
Davis, J.C. (1986) Statistics and Data Analysis in Geology. 2nd Edition, John Wiley & Sons, New York.
}
\keyword{datasets}
