% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectonicr.R
\docType{package}
\name{tectonicr}
\alias{tectonicr}
\alias{_PACKAGE}
\alias{tectonicr-package}
\title{library(tectonicr)}
\description{
A list of documented functions may be viewed by typing
\code{help(package="tectonicr")}. Further details about the theoretical
background are provided by Wdowinski (1999).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/tectonicr/}
  \item Report bugs at \url{https://github.com/tobiste/tectonicr/issues}
}

}
\author{
\strong{Maintainer}: Tobias Stephan \email{tobias.stephan1@yahoo.com} (\href{https://orcid.org/0000-0002-9290-014X}{ORCID})

}
