% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R
\name{transform_matrices}
\alias{transform_matrices}
\alias{tmat1}
\alias{tmat2}
\alias{tmat3}
\title{Transformations into PoR spherical Coordinates}
\usage{
tmat1(x)

tmat2(x)

tmat3(x)
}
\arguments{
\item{x}{angle}
}
\value{
matrix
}
\description{
Helper functions for coordinate transformations
}
