% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{vcross}
\alias{vcross}
\title{Vector cross product}
\usage{
vcross(x, y)
}
\arguments{
\item{x, y}{numeric vectors of length 3}
}
\value{
numeric vector of length 3
}
\description{
Vector or cross product
}
\examples{
vcross(c(1, 2, 3), c(4, 5, 6))
}
