% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vm_qqplot}
\alias{vm_qqplot}
\title{von Mises Quantile-Quantile Plot}
\usage{
vm_qqplot(
  x,
  w = NULL,
  axial = TRUE,
  mean = NULL,
  kappa = NULL,
  xlab = "von Mises quantile function",
  ylab = "Empirical quantile function",
  main = "von Mises Q-Q Plot",
  col = "#B63679FF",
  add_line = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric. Angles in degrees}

\item{w}{numeric. optional weightings for \code{x} to estimate \code{mean} and \code{kappa}.}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})}

\item{mean}{numeric. Circular mean of the von Mises distribution. If \code{NULL},
it will be estimated from \code{x}.}

\item{kappa}{numeric. Concentration parameter of the von Mises distribution.
If \code{NULL}, it will be estimated from \code{x}.}

\item{xlab, ylab, main}{plot labels.}

\item{col}{color for the dots.}

\item{add_line}{logical. Whether to connect the points by straight lines?}

\item{...}{graphical parameters}
}
\value{
plot
}
\description{
Produces a Q-Q plot of the data against a specified von Mises distribution
to graphically assess the goodness of fit of the model.
}
\examples{
set.seed(20250411)

# von Mises distribution
x_vm <- rvm(100, mean = 0, kappa = 2)
vm_qqplot(x_vm, pch = 20)

x_wcauchy <- rwcauchy(100, mean = 0, rho = 0.5)
vm_qqplot(x_wcauchy, pch = 20)

# circular uniform data
x_cunif <- rcunif(100)
vm_qqplot(x_cunif, pch = 20)
}
\seealso{
\code{\link[=vonmises]{vonmises()}}

Other circ-qqplot: 
\code{\link{circular_qqplot}()}
}
\concept{circ-qqplot}
