% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{wcauchy}
\alias{wcauchy}
\alias{rwcauchy}
\alias{dwcauchy}
\alias{pwcauchy}
\alias{qwcauchy}
\title{The Wrapped Cauchy Distribution}
\usage{
rwcauchy(n, mean, rho)

dwcauchy(theta, mean, rho, axial = FALSE, log = FALSE)

pwcauchy(
  theta,
  mean,
  rho,
  axial = FALSE,
  from = NULL,
  lower.tail = TRUE,
  log.p = FALSE
)

qwcauchy(
  p,
  mean,
  rho,
  axial = FALSE,
  from = NULL,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{n}{integer. Number of observations in degrees}

\item{mean}{numeric. The mean vector in degrees.}

\item{rho}{numeric. Concentration parameter in the range (0, 1)}

\item{theta}{numeric. Angular value in degrees}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{log}{logical. If \code{TRUE}, probabilities p are given as \eqn{\log(p)}{log(p)}.}

\item{from}{if \code{NULL} is set to \eqn{\mu-\pi}{mu-pi}. This is the value from
which the \code{pvm} and \code{qvm} are evaluated. in degrees.}

\item{lower.tail}{logical. If \code{TRUE} (default), probabilities are \eqn{P(\Theta \le \theta)},
otherwise \eqn{P(\Theta > \theta)}.}

\item{log.p}{logical. If \code{TRUE}, probabilities p are given as log(p).}

\item{p}{numeric. Vector of probabilities with values in \eqn{[0,1]}{[0,1]}.}
}
\value{
\code{dwcauchy} gives the density,
\code{pwcauchy} gives the probability of the wrapped Cauchy distribution function,
\code{rwcauchy} generates random deviates (in degrees), and
\code{qrwcauchy} provides quantiles (in degrees).
}
\description{
Density, probability distribution function, quantiles, and random generation
for the circular wrapped Cauchy distribution with mean \eqn{\mu} and rho \eqn{\rho}
}
\examples{
set.seed(1)
x <- rwcauchy(5, mean = 90, rho = exp(-1))

dwcauchy(x, mean = 90, rho = exp(-1))
dwcauchy(x, mean = 90, rho = exp(-1), axial = TRUE)

pwcauchy(x, mean = 90, rho = exp(-1))
qwcauchy(c(.25, .5, .75), mean = 90, rho = exp(-1))
}
\seealso{
\link{vonmises} and \link{cunif}
}
