% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{wnorm}
\alias{wnorm}
\alias{rwnorm}
\alias{dwnorm}
\alias{pwnorm}
\alias{qwnorm}
\title{The Wrapped Normal Distribution}
\usage{
rwnorm(n, mean = 0, sd = 1)

dwnorm(theta, mean = 0, sd = 1, axial = FALSE, ..., log = FALSE)

pwnorm(theta, mean = 0, sd = 1, axial = FALSE, from = NULL, ...)

qwnorm(
  p,
  mean = 0,
  sd = 1,
  axial = FALSE,
  from = NULL,
  tol = .Machine$double.eps^(0.6),
  ...
)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{mean}{numeric. The mean vector in degrees.}

\item{sd}{numeric. standard deviation of the (unwrapped) normal distribution in degrees.}

\item{theta}{numeric. Angular value in degrees}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{...}{optional parameters passed to underlying circular functions:
\code{\link[circular:wrappednormal]{circular::dwrappednormal()}}, \code{\link[circular:wrappednormal]{circular::pwrappednormal()}}, and \code{\link[circular:wrappednormal]{circular::qwrappednormal()}}}

\item{log}{logical. If \code{TRUE}, probabilities p are given as \eqn{\log(p)}{log(p)}.}

\item{from}{if \code{NULL} is set to \eqn{\mu-\pi}{mu-pi}. This is the value from
which the \code{pvm} and \code{qvm} are evaluated. in degrees.}

\item{p}{numeric. Vector of probabilities with values in \eqn{[0,1]}{[0,1]}.}

\item{tol}{numeric. The precision in evaluating the distribution function or the quantile.}
}
\value{
\code{dwnorm} gives the density,
\code{pwnorm} gives the probability of the wrapped normal distribution function,
\code{rwnorm} generates random deviates (in degrees), and
\code{qwnorm} provides quantiles (in degrees).
}
\description{
Density, probability distribution function, quantiles, and random generation
for the circular normal distribution with mean \eqn{\mu}{mu} and standard deviation \eqn{\sigma}{sd}.
}
\examples{
set.seed(1)
x <- rwnorm(5, mean = 90, sd = 5)

dwnorm(x, mean = 90, sd = 5, axial = FALSE)
dwnorm(x, mean = 90, sd = 5, axial = TRUE)

pwnorm(x, mean = 90, sd = 5)
qwnorm(c(.25, .5, .75), mean = 90, sd = 5)
}
\seealso{
\link{cunif}, \link{wnorm}, \link{wcauchy}, and \link{vonmises}
}
