% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{wrap_dateline}
\alias{wrap_dateline}
\title{Wrap Spatial object at dateline}
\usage{
wrap_dateline(x)
}
\arguments{
\item{x}{Spatial object}
}
\value{
Spatial object
}
\description{
Wrap a Spatial object at date line
}
