% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsm2016.R
\docType{data}
\name{wsm2016}
\alias{wsm2016}
\title{World Stress Map}
\format{
An object of class \code{"data.frame"}
}
\usage{
data('wsm2016')
}
\description{
The World Stress Map (WSM) is a global compilation of information on the
crustal present-day stress field (Version 2016).
}
\examples{
data("wsm2016")
head("wsm2016")
}
\references{
Heidbach, O., M. Rajabi, X. Cui, K. Fuchs, B. Müller, J.
Reinecker, K. Reiter, M. Tingay, F. Wenzel, F. Xie, M. O. Ziegler,
M.-L. Zoback, and M. D. Zoback (2018): The World Stress Map database
release 2016: Crustal stress pattern across scales. \emph{Tectonophysics}, 744,
484--498, \doi{10.1016/j.tecto.2018.07.007}.
}
\keyword{datasets}
