% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{abs_vel}
\alias{abs_vel}
\title{Absolute plate velocity}
\usage{
abs_vel(w, alpha, r = 6371.00887714)
}
\arguments{
\item{w}{Angular velocity or rate or angle of rotation}

\item{alpha}{Angular distance to Euler pole or small circle around Euler pole}

\item{r}{radius. Default is Earth's radius (6371.00887714 km)}
}
\value{
number
}
\description{
Calculates the absolute angular velocity of plate motion
}
\examples{
abs_vel(0.21, 0)
abs_vel(0.21, 45)
abs_vel(0.21, 90)
}
