% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{angle-conversion}
\alias{angle-conversion}
\alias{rad2deg}
\alias{deg2rad}
\title{Degrees to Radians}
\usage{
rad2deg(rad)

deg2rad(deg)
}
\arguments{
\item{rad}{(array of) angles in radians.}

\item{deg}{(array of) angles in degrees.}
}
\value{
angle in degrees or radians.
}
\description{
Helper functions to transform between angles in degrees and
radians.
}
\examples{
deg2rad(seq(-90, 90, 15))
rad2deg(seq(-pi / 2, pi / 2, length = 13))
}
