% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{angle_vectors}
\alias{angle_vectors}
\title{Angle between two vectors}
\usage{
angle_vectors(x, y)
}
\arguments{
\item{x}{Vector 1 in Cartesian coordinates. Can be a vector of three numbers
or a matrix of 3 columns (first one is x, second y, third z)}

\item{y}{Vector 2. As above}
}
\value{
numeric; angle in degrees
}
\description{
Calculates the angle between two vectors
}
\examples{
u <- c(1, -2, 3)
v <- c(-2, 1, 1)
angle_vectors(u, v)
}
\author{
Tobias Stephan
}
