% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{ar_test}
\alias{ar_test}
\title{Angular Randomisation Test of Homogeneity}
\usage{
ar_test(x, y, n_perm = 1000L, axial = TRUE)
}
\arguments{
\item{x, y}{numeric vectors. Angles in degrees}

\item{n_perm}{integer. Number of permutations}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or circular, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}
}
\value{
list containing the test statistic and the p-value
}
\description{
Performs Angular Randomisation Test for homogeneity on two samples of
circular data after Ruxton et al. (2023).
P-values are estimated using permutation.
}
\examples{
set.seed(20250411)
x1 <- c(35, 45, 50, 55, 60, 70, 85, 95, 105, 120)
x2 <- c(75, 80, 90, 100, 110, 130, 135, 140, 150, 160, 165)
ar_test(x1, x2, axial = FALSE)

# San Andreas Fault Data:
data(san_andreas)
data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
ar_test(sa.por$azi.PoR, rvm(100, 135, 10))
}
\references{
Ruxton, G.D., Malkemper, E.P. & Landler, L. Evaluating the power
of a recent method for comparing two circular distributions: an alternative
to the Watson U2 test. Sci Rep 13, 10007 (2023). https://doi.org/10.1038/s41598-023-36960-1
}
\seealso{
Other Tests: 
\code{\link{kuiper_test}()},
\code{\link{norm_chisq}()},
\code{\link{rayleigh-test}},
\code{\link{watson_test}()},
\code{\link{watson_two_sample}},
\code{\link{watson_wheeler_test_perm}()},
\code{\link{weighted-rayleigh-test}}
}
\concept{Tests}
