% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circle_median}
\alias{circle_median}
\alias{circular_quasi_median}
\alias{circular_mean}
\alias{circular_quasi_quantile}
\alias{circular_quasi_IQR}
\alias{circular_var}
\alias{circular_mean_deviation}
\alias{circular_median_deviation}
\alias{circular_mean_error}
\title{Median and statistics on Pi-periodic Data}
\source{
\code{\link[=median]{median()}}, \code{\link[=quantile]{quantile()}}, and \code{\link[=IQR]{IQR()}} are the
equivalents for non-periodic data.
}
\usage{
circular_quasi_median(x, na.rm = TRUE)

circular_mean(x, na.rm = TRUE)

circular_quasi_quantile(x, na.rm = TRUE)

circular_quasi_IQR(x, na.rm = TRUE)

circular_var(x, na.rm = TRUE)

circular_mean_deviation(x, na.rm = TRUE)

circular_median_deviation(x, na.rm = TRUE)

circular_mean_error(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector in degrees.}

\item{na.rm}{logical. Should missing values (including \code{NaN}) be removed?}
}
\value{
Numeric vector
}
\description{
Calculate the mean, median, quartile, interquartile range,
variance, deviation, and error of directional data.
}
\details{
Quasi median on the circle, quasi quartiles on a circle, quasi
interquartile range on a circle.
}
\examples{
x <- c(0, 45, 55, 40 + 180, 50 + 180, NA)
circular_mean(x)
circular_quasi_median(x)
circular_quasi_quantile(x)
circular_quasi_IQR(x)
circular_var(x)
circular_mean_deviation(x)
circular_median_deviation(x)
circular_mean_error(x)

data("san_andreas")
circular_quasi_median(san_andreas$azi)
}
\references{
\itemize{
\item Ratanaruamkarn, S., Niewiadomska-Bugaj, M., Wang, J.-C. (2009).
A New Estimator of a Circular Median. \emph{Communications in Statistics -
Simulation and Computation}, \strong{38}(6), 1269-1291.
\doi{10.1080/03610910902899950}.
\item Reiter, K., Heidbach, O., Schmitt, D., Haug, K., Ziegler, M., & Moeck, I.
(2014). A revised crustal stress orientation database for Canada.
\emph{Tectonophysics}, \strong{636}, 111-124. \doi{10.1016/j.tecto.2014.08.006}
}
}
