% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{circle_median}
\alias{circle_median}
\alias{circular_quasi_median}
\alias{circular_quasi_quartile}
\alias{circular_quasi_interquartile_range}
\alias{circular_variance}
\alias{circular_mean_deviation}
\alias{circular_median_deviation}
\alias{circular_mean_error}
\title{Median and Quartile on Pi-periodic Data}
\source{
\code{\link[stats:median]{stats::median()}}, \code{\link[stats:stats-package]{stats::stats()}}, and \code{\link[stats:IQR]{stats::IQR()}} are the
equivalents for non-periodic data.
}
\usage{
circular_quasi_median(x, quiet = TRUE)

circular_quasi_quartile(x, quiet = TRUE)

circular_quasi_interquartile_range(x, quiet = TRUE)

circular_variance(x, quiet = TRUE)

circular_mean_deviation(x, quiet = TRUE)

circular_median_deviation(x, quiet = TRUE)

circular_mean_error(x, quiet = TRUE)
}
\arguments{
\item{x}{Numeric vector (NA values will be removed).}

\item{quiet}{logical. If false, a warning message is printed if there are NA
values.}
}
\value{
Numeric vector
}
\description{
Calculate the median, quartile, and interquartile range of
orientation data.
}
\details{
Quasi median on the circle, quasi quartiles on a circle, quasi
interquartile range on a circle.
}
\examples{
x <- c(0, 45, 55, 40 + 180, 50 + 180)
circular_quasi_median(x)
circular_quasi_quartile(x)
circular_quasi_interquartile_range(x)
circular_variance(x)
circular_mean_deviation(x)
circular_median_deviation(x)
circular_mean_error(x)
}
\references{
\itemize{
\item Ratanaruamkarn, S., Niewiadomska-Bugaj, M., Wang, J.-C. (2009).
A New Estimator of a Circular Median. \emph{Communications in Statistics -
Simulation and Computation}, 38(6), 1269--1291.
\doi{10.1080/03610910902899950}.
}
}
