% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{circular_density}
\alias{circular_density}
\title{Circular Kernel Density}
\usage{
circular_density(
  x,
  z = NULL,
  bw = NULL,
  na.rm = TRUE,
  from = 0,
  to = 360,
  n = 512,
  axial = TRUE,
  kappa = NULL,
  rho = NULL,
  kernel = c("vonmises", "wrappedcauchy"),
  adjust = 1
)
}
\arguments{
\item{x}{numeric. A vector of angles (in degrees) from which the estimate is to be computed.}

\item{z}{numeric. Angles where the density is estimated. If \code{NULL} equally
spaced angles are used according to the parameters \code{from}, \code{to} and \code{n}.}

\item{bw, kappa, rho}{numeric. Smoothing bandwidth expressed as the concentration
parameter \eqn{\kappa}{k} for the von Mises distribution or \eqn{\rho}{rho} for the
wrapped Cauchy distribution.
Small and large values for the von Mises and wrapped Cauchy distribution,
respectively, gives smooth density lines. If not specified, parameter will be estimated using
\code{\link[=est.kappa]{est.kappa()}}  for the von Mises distribution, or set to \eqn{p \exp(-1)}{p exp(-1)}
for the wrapped Cauchy distribution (where \eqn{p = 2}{p=2} when \code{axial=TRUE} and 1 otherwise).}

\item{na.rm}{logical; if \code{TRUE}, missing values are removed
    from \code{x}. If \code{FALSE} any missing values cause an error.}

\item{from, to}{the left and right-most points of the grid at which the
    density is to be estimated; the defaults are \code{cut * bw} outside
    of \code{range(x)}.}

\item{n}{integer. Number of equally spaced angles at which the density is
to be estimated.}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})
or biaxial (\code{TRUE}, the default).}

\item{kernel}{character. The smoothing kernel to be used; one of \code{"vonmises"}
(the default) or \code{"wrappedcauchy"} for the von Mises or the Wrapped Cauchy
distribution.}

\item{adjust}{the bandwidth used is actually \code{adjust*bw}.
    This makes it easy to specify values like \sQuote{half the default}
    bandwidth.}
}
\value{
Object of class \code{"density"}
}
\description{
Multiples of a von Mises density or wrapped Cauchy distribution for circular data
}
\examples{
# von Mises kernel density
circular_density(san_andreas$azi, kappa = 100)

# wrapped Cauchy kernel density
circular_density(san_andreas$azi, rho = 0.9, kernel = "wrappedcauchy")
}
\seealso{
\code{\link[stats:density]{stats::density()}}, \code{\link[=dvm]{dvm()}} and \code{\link[=dwcauchy]{dwcauchy()}}
}
