% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{circular_qqplot}
\alias{circular_qqplot}
\title{Quantile-Quantile Linearised Plot for Circular Distributions}
\usage{
circular_qqplot(
  x,
  axial = TRUE,
  xlab = paste("i/(n+1)"),
  ylab = NULL,
  main = "Circular Quantile-Quantile Plot",
  add_line = TRUE,
  col = "#B63679FF",
  ...
)
}
\arguments{
\item{x}{numeric. Angles in degrees}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})}

\item{xlab, ylab, main}{plot labels.}

\item{add_line}{logical. Whether to connect the points by straight lines?}

\item{col}{color for the dots.}

\item{...}{graphical parameters}
}
\value{
plot
}
\description{
Uniformly distributed orientations should yield a straight line through the
origin. Systematic departures from linearity will indicate preferred
orientation.
}
\examples{
set.seed(20250411)

# von Mises distribution
x_vm <- rvm(100, mean = 0, kappa = 2)
circular_qqplot(x_vm, pch = 20)

x_wcauchy <- rwcauchy(100, mean = 0, rho = 0.5)
circular_qqplot(x_wcauchy, pch = 20)

# circular uniform data
x_cunif <- rcunif(100)
circular_qqplot(x_cunif, pch = 20)
}
\references{
Borradaile, G. J. (2003). Statistics of earth
science data: their distribution in time, space, and orientation (Vol. 351,
p. 329). Berlin: Springer.
}
\seealso{
Other circ-qqplot: 
\code{\link{vm_qqplot}()}
}
\concept{circ-qqplot}
