% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{compact_grid}
\alias{compact_grid}
\title{Compact smoothed stress field}
\usage{
compact_grid(x, type = c("stress2grid", "dispersion_grid"))
}
\arguments{
\item{x}{output of \code{\link[=stress2grid]{stress2grid()}} or \code{\link[=PoR_stress2grid]{PoR_stress2grid()}}}

\item{type}{type of the grid. Either "stress2grid" of "dispersion_grid"}
}
\value{
\code{sf} object
}
\description{
Filter smoothed stress field to smallest wavelength (R) for each coordinate
}
\examples{
data("san_andreas")
res <- stress2grid(san_andreas)
compact_grid(res)
}
