% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{compact_grid}
\alias{compact_grid}
\title{Compact smoothed stress field}
\usage{
compact_grid(x)
}
\arguments{
\item{x}{output of \code{stress2grid()} or \code{PoR_stress2grid()}}
}
\value{
\code{sf} object
}
\description{
Filter smoothed stress field to lowest R per coordinate
}
\examples{
data("san_andreas")
res <- stress2grid(san_andreas)
compact_grid(res)
}
