% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{data2PoR}
\alias{data2PoR}
\title{Transforms coordinates and azimuths into PoR coordinates system}
\usage{
data2PoR(x, PoR)
}
\value{
\code{sf} object in PoR CRS with additional columns \code{lon.PoR},
\code{lat.PoR}, and \code{azi.PoR}
}
\description{
Convenience function to add PoR coordinates and PoR azimuths to data
}
\examples{
por <- subset(nuvel1, nuvel1$plate.rot == "na")
data2PoR(san_andreas, por)
}
