% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation_norm.R
\name{deviation_norm}
\alias{deviation_norm}
\title{Normalize angle between two directions}
\usage{
deviation_norm(x)
}
\arguments{
\item{x}{numeric vector containing angles in degrees}
}
\value{
numeric vector, acute angles between two directions, i.e. values
between 0° and 90°
}
\description{
normalizes the angle between two directions to the acute angle in between,
i.e. angles between 0 and 90 degrees
}
\examples{

}
\author{
Copyright (C) 2021 Tobias Stephan
}
