% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dispersion}
\alias{dispersion}
\alias{circular_distance}
\alias{circular_dispersion}
\title{Circular distance and dispersion}
\usage{
circular_distance(x, y, axial = TRUE, na.rm = TRUE)

circular_dispersion(x, from = NULL, w = NULL, axial = TRUE, na.rm = TRUE)
}
\arguments{
\item{x, y, from}{vector of numeric values in degrees.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(\code{TRUE}, the default) or circular, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}
}
\description{
Circular distance between two angles and circular dispersion of angles
about a specified angle.
}
\examples{
a <- c(0, 2, 359, 6, 354)
b <- a + 20
circular_distance(a, b)

data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, ep, "right")
circular_dispersion(sa.por$azi.PoR, from = 135)
}
\seealso{
\code{\link[=circular_mean]{circular_mean()}}
}
