% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{dispersion}
\alias{dispersion}
\alias{circular_distance}
\alias{circular_dispersion}
\title{Circular distance and dispersion}
\usage{
circular_distance(x, y, axial = TRUE, na.rm = TRUE)

circular_dispersion(
  x,
  mean = NULL,
  w = NULL,
  norm = FALSE,
  axial = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{x, y}{vectors of numeric values in degrees.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{na.rm}{logical. Whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}

\item{mean}{numeric. The angle about which the angles \code{x} disperse (in degrees).}

\item{w}{(optional) Weights. A vector of positive numbers and of the same
length as \code{x}.}

\item{norm}{logical. Whether the dispersion should be normalized by the
maximum possible angular difference.}
}
\value{
numeric.
}
\description{
Circular distance between two angles and circular dispersion of angles
about a specified angle.
}
\note{
If \code{from} is \code{NULL}, than the circular variance is returned.
}
\examples{
a <- c(0, 2, 359, 6, 354)
b <- a + 90
circular_distance(a, b)

data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
circular_dispersion(sa.por$azi.PoR, mean = 135)
circular_dispersion(sa.por$azi.PoR, mean = 135, w = 1 / san_andreas$unc)
}
\references{
Mardia, K.V. (1972). Statistics of Directional Data: Probability
and Mathematical Statistics. London: Academic Press.
}
\seealso{
\code{\link[=circular_mean]{circular_mean()}}, \code{\link[=circular_var]{circular_var()}}.
}
