% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{dist_greatcircle}
\alias{dist_greatcircle}
\title{Distance between points}
\usage{
dist_greatcircle(
  lats,
  lons,
  point_lat,
  point_lon,
  r = earth_radius(),
  method = c("haversine", "haversine2", "vincenty")
)
}
\arguments{
\item{point_lat, point_lon}{coordinates of points (degrees). Must be of
length = 1.}

\item{r}{radius of the sphere (default = 6371.0087714 km, i.e. the radius of
the Earth)}

\item{method}{Formula for calculating great circle distance:
\describe{
\item{"haversine"}{Haversine formula (the default)}
\item{"haversine2"}{Haversine formula optimized for 64-bit floating-point numbers}
\item{"vincenty"}{Vincenty formula for an ellipsoid with equal major and minor axes}
}}

\item{grid_lat, grid_lon}{coordinate of grid point (degrees).}
}
\description{
Returns the great-circle distance between a location and all grid point in km
}
\examples{
dist_greatcircle(grid_lat = 20, gird_lon = 12, point_lat = c(50, 30),
point_lon = c(40, 32))
dist_greatcircle(grid_lat = 20, gird_lon = 12, point_lat = c(50, 30),
point_lon = c(40, 32), method = "haversine2")
dist_greatcircle(grid_lat = 20, gird_lon = 12, point_lat = c(50, 30),
point_lon = c(40, 32), method = "vincenty")
}
\seealso{
\code{\link[=orthodrome_haversine]{orthodrome_haversine()}}, \code{\link[=orthodrome_haversine2]{orthodrome_haversine2()}}, \code{\link[=orthodrome_vincenty]{orthodrome_vincenty()}}
}
