% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{distance_from_pb}
\alias{distance_from_pb}
\title{Distance from plate boundary}
\usage{
distance_from_pb(x, ep, pb, tangential = FALSE)
}
\arguments{
\item{x, pb}{\code{sf} objects of the data points and the plate boundary
geometries in the geographical coordinate system}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole
(\code{lat}, \code{lon})}

\item{tangential}{Logical. \code{TRUE} for tangential boundaries and
\code{FALSE} (the default) for inward and outward boundaries.}
}
\value{
Numeric vector of the great circle distances in degree
}
\description{
Absolute distance of data points from the nearest plate boundary in degree
}
\details{
The distance to the plate boundary is the longitudinal or
latitudinal difference between the data point and the plate boundary
(along the closest latitude or longitude) for inward/outward or tangential
plate boundaries, respectively.
}
\examples{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("PB2002")
plate_boundary <- subset(
  PB2002, PB2002$PlateA \%in\% c("NA", "PA") & PB2002$PlateB \%in\% c("NA", "PA")
)

data("san_andreas")
res <- distance_from_pb(
  x = san_andreas, ep = na_pa, pb = plate_boundary, tangential = TRUE
)
head(res)
}
