% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_distance.R
\name{distance_from_pb}
\alias{distance_from_pb}
\title{Distance from plate boundary}
\usage{
distance_from_pb(x, euler, pb, tangential = FALSE, km = FALSE, ...)
}
\arguments{
\item{x, pb}{\code{sf} objects of the data points and the plate boundary
geometries in the geographical coordinate system}

\item{euler}{\code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Pole of Rotation}

\item{tangential}{Logical. Whether the plate boundary is a tangential
boundary (\code{TRUE}) or an inward and outward boundary (\code{FALSE}, the
default).}

\item{km}{Logical. Whether the distance is expressed in kilometers
(\code{TRUE}) or in degrees (\code{FALSE}, the default).}

\item{...}{optional arguments passed to \code{\link[smoothr:densify]{smoothr::densify()}}}
}
\value{
Numeric vector of the great circle distances
}
\description{
Absolute distance of data points from the nearest plate boundary in degree
}
\details{
The distance to the plate boundary is the longitudinal or
latitudinal difference between the data point and the plate boundary
(along the closest latitude or longitude) for inward/outward or tangential
plate boundaries, respectively.
}
\examples{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")

data("san_andreas")
res <- distance_from_pb(
  x = san_andreas, euler = na_pa, pb = plate_boundary, tangential = TRUE
)
head(res)

res.km <- distance_from_pb(
  x = san_andreas, euler = na_pa, pb = plate_boundary, tangential = TRUE, km = TRUE
)
range(res.km)
}
