% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{dummy}
\alias{dummy}
\alias{smallcircle_dummy}
\alias{greatcircle_dummy}
\alias{loxodrome_dummy}
\title{Plate Stress Dummy Grid}
\usage{
smallcircle_dummy(n)

greatcircle_dummy(n)

loxodrome_dummy(n, angle, sense)
}
\arguments{
\item{n}{Number of curves}

\item{angle}{Direction of loxodromes (in degree)}

\item{sense}{Sense of loxodromes 'sinistral' or 'dextral' for 'clockwise' or
'counterclockwise' loxodromes, respectively}
}
\value{
\code{data.frame}
}
\description{
Creates a dummy grid for small circles, great circles, and
loxodromes of an Euler pole
}
\examples{
smallcircle_dummy(10)
greatcircle_dummy(10)
loxodrome_dummy(10, 45, "sinistral")
}
