% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{eulerpole_greatcircles}
\alias{eulerpole_greatcircles}
\title{Draws great-circles of Euler pole}
\usage{
eulerpole_greatcircles(x, d, gridsize)
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler poles in lat, lon, and rotation angle (optional)}

\item{d}{numeric; number of equally spaced great circles (angle between great circles (number of great circles = 360 / angle); default = 12}

\item{gridsize}{numeric, angle between great circles}
}
\value{
An object of class \code{SpatialLinesDataFrame}
}
\description{
Calculate the direction of maximum horizontal stress along great circles,
small circles, and loxodromes around the relative plate motion´s Euler pole
at a given point or points
}
\details{
if angle is given: output additionally gives absolute velocity on small-circle (degree/Myr -> km/Myr)
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
euler$angle <- euler$rate
eulerpole_greatcircles(euler)
}
\author{
Tobias Stephan
}
