% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{eulerpole_greatcircles}
\alias{eulerpole_greatcircles}
\title{Draws great-circles of Euler pole}
\usage{
eulerpole_greatcircles(x, gm, n)
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler poles in lat, lon, and rotation angle (optional)}

\item{gm}{numeric, angle between great circles}

\item{n}{numeric; number of equally spaced great circles (angle between great circles (number of great circles n = 360 / gm); default = 12}
}
\value{
An object of class \code{SpatialLinesDataFrame}
}
\description{
.
}
\details{
if angle is given: output additionally gives absolute velocity on small-circle (degree/Myr -> km/Myr)
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
euler$angle <- euler$rate
eulerpole_greatcircles(euler)
}
\author{
Tobias Stephan
}
