% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{eulerpole_smallcircles}
\alias{eulerpole_smallcircles}
\title{Small-circles around Euler pole}
\usage{
eulerpole_smallcircles(x, sm, sf = TRUE)
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler pole in lat, lon,
and rotation angle (optional)}

\item{sm}{numeric, angle between small circles (in degree); default: 10}

\item{sf}{logical. Export object type of the lines.
TRUE (the default) simple feature (\code{sf}) objects.
FALSE for \code{"SpatialLInesDataFrame"} object.}
}
\value{
An object of class \code{sf} or \code{SpatialLinesDataFrame}
}
\description{
Construct a line that has a constant distance to the Euler pole.
}
\details{
if angle is given: output additionally gives absolute velocity on
small circle (degree/Myr -> km/Myr)
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
euler$angle <- euler$rate
# Pacific plate
eulerpole_smallcircles(euler)
}
\author{
Tobias Stephan
}
