% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{eulerpole_smallcircles}
\alias{eulerpole_smallcircles}
\title{Draws small-circles around Euler pole}
\usage{
eulerpole_smallcircles(x, gridsize = 10)
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler pole in lat, lon (and rotation angle \link{optional})}

\item{gridsize}{numeric, angular distance between small circles; default: 10}
}
\value{
An object of class \code{SpatialLinesDataFrame}
}
\description{
Calculate the direction of maximum horizontal stress along great circles,
small circles, and loxodromes around the relative plate motion´s Euler pole
at a given point or points
}
\details{
if angle is given: output additionally gives absolute velocity on small-circle (degree/Myr -> km/Myr)
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
euler$angle <- euler$rate
# Pacific plate
eulerpole_smallcircles(euler)
}
\author{
Tobias Stephan
}
