% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_azimuth.R
\name{get_azimuth}
\alias{get_azimuth}
\title{Azimuth between two points}
\usage{
get_azimuth(p, q)
}
\arguments{
\item{p1}{two-column vector containing latitudes and longitudes of start point}

\item{p2}{two-column vector containing latitudes and longitudes of end point}
}
\value{
azimuth between two points at start point p1
}
\description{
Calculate initial bearing (or forward azimuth) which is followed along a
straight line along a great-circle arc between the start point and the end
point
}
\examples{
p1 <- c(35, 45) # Baghdad
p2 <- c(35, 135) # Osaka
get_azimuth(p1, p2)

}
\references{
http://www.movable-type.co.uk/scripts/latlong.html
}
\author{
Copyright (C) 2021 Tobias Stephan
}
