% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{greatcircle_dummy}
\alias{greatcircle_dummy}
\title{Great circle grid dummy}
\usage{
greatcircle_dummy(x)
}
\arguments{
\item{x}{number of great circles}
}
\value{
data.frame
}
\description{
Construct curves cutting small circles perpendicular.
}
