% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stress.R
\docType{data}
\name{homing}
\alias{homing}
\title{Homing directions of Pidgeons}
\format{
numeric vector of length 10

numeric vector of length 51
}
\description{
Directions of 10 pidgeons, example data from Mardia and Jupp (1999), pp. 93

Directions of 51 glacial striations, example data from Davis (1986), pp. 316
}
\examples{
data("homing")
homing
data("striae")
striae
}
\references{
Mardia and Jupp (1999). Directional Statistics. John Wiley and
Sons.

Davis, J.C. (1986) Statistics and Data Analysis in Geology. 2nd Edition, John Wiley & Sons, New York.
}
\keyword{datasets}
