% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{is.euler}
\alias{is.euler}
\title{Check if object is euler.pole}
\usage{
is.euler(x)
}
\arguments{
\item{x}{object of class \code{"euler.pole"}}
}
\value{
logical
}
\description{
Check if object is euler.pole
}
