% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{kernel_dispersion}
\alias{kernel_dispersion}
\alias{dispersion_grid}
\title{Adaptive Kernel Dispersion}
\usage{
kernel_dispersion(
  x,
  stat = c("dispersion", "nchisq", "rayleigh"),
  grid = NULL,
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  min_data = 3L,
  max_data = Inf,
  min_dist_threshold = 200,
  dist_threshold = 0.1,
  stat_threshold = Inf,
  R_range = seq(100, 2000, 100),
  ...
)

dispersion_grid(...)
}
\arguments{
\item{x}{\code{sf} object containing
\describe{
\item{azi}{\eqn{\sigma_\text{Hmax}}{SHmax} in degree}
\item{unc}{(optional) Uncertainties of SHmax in degree}
\item{type}{(optional) Methods used for the determination of the direction
of \eqn{\sigma_\text{Hmax}}{SHmax}}
}}

\item{stat}{The measurement of dispersion to be calculated. Either
\code{"dispersion"} (default), \code{"nchisq"}, or \code{"rayleigh"} for circular
dispersion, normalized Chi-squared test statistic, or Rayleigh test
statistic.}

\item{grid}{(optional) Point object of class \code{sf}.}

\item{lon_range, lat_range}{(optional) numeric vector specifying the minimum
and maximum longitudes and latitudes (ignored if \code{grid} is specified).}

\item{gridsize}{numeric. Target spacing of the regular grid in decimal
degree. Default is \code{2.5}. (is ignored if \code{grid} is specified)}

\item{min_data}{integer. If the number of observations within distance
\code{R_range} is less than \code{min_data}, a missing value \code{NA} will be generated.
Default is \code{3} for \code{\link[=stress2grid]{stress2grid()}} and \code{4} for \code{\link[=stress2grid_stats]{stress2grid_stats()}}.}

\item{max_data}{integer. The number of nearest observations that should be
used for prediction, where "nearest" is defined in terms of the space of the
spatial locations. Default is \code{Inf}.}

\item{min_dist_threshold}{numeric. Distance threshold for smallest distance
of the prediction location to the next observation location.
Default is \code{200} km.}

\item{dist_threshold}{numeric. Distance weight to prevent overweight of data
nearby (0 to 1). Default is \code{0.1}}

\item{stat_threshold}{numeric. Generates missing values when the kernel
\code{stat} value exceeds this threshold. Default is \code{Inf}.}

\item{R_range}{numeric value or vector specifying the kernel half-width(s)
search radii,
i.e. the maximum distance from the prediction location to be used for
prediction (in km). Default is \code{seq(50, 1000, 50)}. If combined with
\code{max_data}, both criteria apply.}

\item{...}{(optional) arguments to \code{\link[=dist_greatcircle]{dist_greatcircle()}}}
}
\value{
\code{sf} object containing
\describe{
\item{lon,lat}{longitude and latitude in degree}
\item{stat}{output of function defined in \code{stat}}
\item{R}{The rearch radius in km.}
\item{mdr}{Mean distance of datapoints per search radius}
\item{N}{Number of data points in search radius}
}
}
\description{
Stress field and wavelength analysis using circular dispersion
(or other statistical estimators for dispersion)
}
\note{
\code{dispersion_grid()} was renamed to \code{kernel_dispersion()} to create
a more consistent API.
}
\examples{
data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
san_andreas_por <- data2PoR(san_andreas, PoR)
san_andreas_por$prd <- 135
kernel_dispersion(san_andreas_por) |> head()
}
\seealso{
\code{\link[=circular_dispersion]{circular_dispersion()}}, \code{\link[=norm_chisq]{norm_chisq()}}, \code{\link[=rayleigh_test]{rayleigh_test()}}
}
\keyword{internal}
