% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{longitude_modulo}
\alias{longitude_modulo}
\title{Longitude correction}
\usage{
longitude_modulo(longitude)
}
\arguments{
\item{longitude}{longitude(s) in degree}
}
\value{
number
}
\description{
Corrects the longitude value to values between -180 and +180
degree
}
\examples{
longitude_modulo(-361)
}
