% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{mean_SC}
\alias{mean_SC}
\title{Mean cosine and sine}
\usage{
mean_SC(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{angles in degrees}

\item{w}{weightings}

\item{na.rm}{logical}
}
\value{
named two element vector
}
\description{
Mean cosine and sine
}
\examples{
\dontrun{
x <- rvm(100, 0, 5)
mean_SC(x)
}
}
