% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{model_shmax}
\alias{model_shmax}
\title{Modeling directions of maximum horizontal stress}
\usage{
model_shmax(df, euler)
}
\arguments{
\item{df}{\code{data.frame} containing the coordinates of the point(s)
(lat, lon)}

\item{euler}{\code{data.frame} containing the coordinates of the Euler pole
for the plate boundary (lat, lon)}
}
\value{
An object of class \code{data.frame}
\describe{
\item{gc}{azimuth of the modeled maximum horizontal following stress
great circles}
\item{sc}{small circles}
\item{ld.cw}{clockwise loxodromes}
\item{ld.ccw}{counter-clockwise loxodromes}
}
}
\description{
Calculate the direction of maximum horizontal stress along great circles,
small circles, and loxodromes around the relative plate motion´s Euler pole
at a given point or points
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
model_shmax(point, euler)
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, http://dx.doi.org/10.1029/97JB03390.
}
\author{
Tobias Stephan
}
