% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{model_shmax}
\alias{model_shmax}
\title{Modeling directions of maximum horizontal stress}
\usage{
model_shmax(df, euler)
}
\arguments{
\item{df}{data.frame containing the coordinates of the points (lat, lon)}

\item{euler}{data.frame containing the coordinates of the Euler pole for the
responsible plate boundary (lat, lon)}
}
\value{
data.frame containing the azimuth of the modeled maximum horizontal
stress great circles (gc), small circles (sc), and loxodromes (ld.cw, ld.ccw)
at the given point
}
\description{
Calculate the direction of maximum horizontal stress along great circles,
small circles, and loxodromes around the relative plate motion´s Euler pole
at a given point or points
}
\examples{
point <- data.frame(lat = 45, lon = 20)
euler <- data.frame(lat = 90, lon = 0)
plate_vector(point, euler)
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, http://dx.doi.org/10.1029/97JB03390.
}
\author{
Copyright (C) 2021 Tobias Stephan
}
