% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{model_shmax}
\alias{model_shmax}
\title{Theoretical Direction of Maximum Horizontal Stress}
\usage{
model_shmax(df, euler)
}
\arguments{
\item{df}{\code{data.frame} containing the coordinates of the point(s)
(lat, lon)}

\item{euler}{\code{data.frame} containing the coordinates of the Euler pole
for the plate boundary (lat, lon)}
}
\value{
\code{data.frame}
\describe{
\item{gc}{Azimuth of modeled \eqn{\sigma_\text{Hmax}}{SHmax} following
great circles}
\item{sc}{Small circles}
\item{ld.cw}{Clockwise loxodromes}
\item{ld.ccw}{Counter-clockwise loxodromes}
}
}
\description{
Models the direction of maximum horizontal stress \eqn{\sigma_\text{Hmax}}{SHmax} along
great circles, small circles, and loxodromes at a given point or points
according to the relative plate motion
}
\details{
\eqn{\sigma_\text{Hmax}}{SHmax} following \emph{great circles} is the
(initial) bearing between the given point and the pole of relative plate
motion. \eqn{\sigma_\text{Hmax}}{SHmax} along \emph{small circles}, clockwise, and
counter-clockwise \emph{loxodromes} is 90\eqn{^{\circ}}{ degree}, +45\eqn{^{\circ}}{ degree}, and 135\eqn{^{\circ}}{ degree} (-45\eqn{^{\circ}}{ degree})
to this great circle bearing, respectively.
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
model_shmax(point, euler)
}
\seealso{
\code{\link[=misfit_shmax]{misfit_shmax()}} to compute the deviation of the modeled direction
from the observed direction of \eqn{\sigma_\text{Hmax}}{SHmax}.
}
\author{
Tobias Stephan
}
