% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_chi2.R
\name{norm_chi2}
\alias{norm_chi2}
\title{Normalized chi-square test}
\usage{
norm_chi2(obs, prd, unc)
}
\arguments{
\item{obs}{observed SHmax, numeric vector}

\item{prd}{predicted SHmax, numeric vector of length of obs}

\item{unc}{uncertainty of observed SHmax, either numeric vector of length of
obs or a number}
}
\value{
numeric vector
}
\description{
A quantitative comparison between the predicted and observed directions of
SHmax is obtained by the calculation of the average azimuth and by a
normalized chi-square test.
}
\details{
Test values are between 0 and 1 indicating the quality of the
predicted SHmax directions. Low values (<= 0.15) indicate good agreement,
high values (>0.7) indicate a systematic misfit between predicted and
observed SHmax directions
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID=='na') #North America relative to Pacific plate
point <- data.frame(lat = 45, lon = 20)
prd <- model_shmax(point, euler)
norm_chi2(obs=90, prd, unc=10)

}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, http://dx.doi.org/10.1029/97JB03390.
}
\author{
Copyright (C) 2021 Tobias Stephan
}
