% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{norm_chi2}
\alias{norm_chi2}
\title{Normalized chi-square test}
\usage{
norm_chi2(obs, prd, unc)
}
\arguments{
\item{obs}{numeric vector containing the observed azimuth of SHmax,
same length as \code{prd}}

\item{prd}{\code{data.frame} containing the modeled azimuths of SHmax, i.e.
the return object from \code{model_shmax()}}

\item{unc}{uncertainty of observed SHmax, either numeric vector of length of
\code{obs} or a number}
}
\value{
numeric vector
}
\description{
A quantitative comparison between the predicted and observed directions of
SHmax is obtained by the calculation of the average azimuth and by a
normalized chi-square test.
}
\details{
Test result are values are between 0 and 1 indicating the quality of
the predicted SHmax directions. Low values (<= 0.15) indicate good agreement,
high values (>0.7) indicate a systematic misfit between predicted and
observed SHmax directions
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
prd <- model_shmax(point, euler)
norm_chi2(obs = 90, prd$sc, unc = 10)
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics: Journal of Geophysical Research: Solid Earth, v. 103, p.
5037-5059, http://dx.doi.org/10.1029/97JB03390.
}
\author{
2021 Tobias Stephan
}
