% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{norm_chi2}
\alias{norm_chi2}
\title{Normalized Chi-Square Test}
\usage{
norm_chi2(obs, prd, unc)
}
\arguments{
\item{obs}{Numeric vector containing the observed azimuth of \eqn{\sigma_\text{Hmax}}{SHmax},
same length as \code{prd}}

\item{prd}{\code{data.frame} containing the modeled azimuths of \eqn{\sigma_\text{Hmax}}{SHmax}, i.e.
the return object from \code{model_shmax()}}

\item{unc}{Uncertainty of observed \eqn{\sigma_\text{Hmax}}{SHmax}, either
numeric vector of length of
\code{obs} or a number}
}
\value{
Numeric vector
}
\description{
A quantitative comparison between the predicted and observed directions of
\eqn{\sigma_\text{Hmax}}{SHmax} is obtained by the calculation of the average
azimuth and by a
normalized \eqn{\chi^2}{chi-square} test.
}
\details{
The normalized \eqn{\chi^2}{chi-square} test is
\deqn{ \text{Norm} \chi^2_i =
 = \frac{
   \sum^M_{i = 1} \left( \frac{\alpha_i - \alpha_{\text{predict}}}{\sigma_i} \right) ^2}
   {\sum^M_{i = 1} \left( \frac{90}{\sigma_i} \right) ^2 }}{
   (sum( ((obs-prd)/unc)^2 ) / sum( (90/unc)^2 )
   }
The test result are values are between 0 and 1 indicating the quality of
the predicted \eqn{\sigma_\text{Hmax}}{SHmax} directions. Low values
(\eqn{\le 0.15}) indicate good agreement,
high values (\eqn{> 0.7}) indicate a systematic misfit between predicted and
observed \eqn{\sigma_\text{Hmax}}{SHmax} directions.
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
prd <- model_shmax(point, euler)
norm_chi2(obs = 90, prd$sc, unc = 10)
}
\references{
Wdowinski, S., 1998, A theory of intraplate
tectonics. \emph{Journal of Geophysical Research: Solid Earth}, v. 103, p.
5037--5059, \doi{10.1029/97JB03390}.
}
