% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ort-eigen}
\alias{ort-eigen}
\alias{ot_eigen2d}
\alias{principal_direction}
\alias{orientation_strength}
\title{Decomposition of Orientation Tensor}
\usage{
ot_eigen2d(x, w = NULL, axial = TRUE, scale = FALSE)

principal_direction(x, w = NULL, axial = TRUE)

orientation_strength(x, w = NULL, axial = TRUE)
}
\arguments{
\item{x}{numeric vector. Values in degrees.}

\item{w}{(optional) Weights. A vector of positive numbers and of the same
length as \code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{scale}{logical. Whether the Eigenvalues should be scaled so they sum
up to 1. Only applicable when weighting are specified,
otherwise the eigenvalues are always scaled.}
}
\value{
\code{ot_eigen2d} returns a list of the Eigenvalues and the angles corresponding to the Eigenvectors.
\code{principal_direction()} and \code{orientation_strength()} are convenience functions
to return the orientation of the largest eigenvalue, and the orientation strength, respectively.
}
\description{
Eigenvector decomposition of the orientation tensor
}
\details{
The two perpendicular \strong{Eigenvectors} are the "principal directions" towards the
highest and the lowest concentration of orientation data.

The \strong{Eigenvalues} can be
interpreted as the fractions of the data explained by the
orientation of the associated principal direction.
Thus, the strength of the orientation is the largest eigenvalue normalized
by the sum of the eigenvalues (\code{scale=TRUE}).
}
\note{
Eigenvalues and eigenvectors of the orientation tensor (inertia tensor) are also called
"principle moments of inertia"  and "principle axes of inertia", respectively.
}
\examples{
test <- rvm(100, mean = 0, k = 10)
ot_eigen2d(test, axial = FALSE)

data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
sa_eig <- ot_eigen2d(sa.por$azi.PoR, w = weighting(san_andreas$unc), scale = TRUE)
print(sa_eig)

rose(sa.por$azi.PoR, muci = FALSE)
rose_line(sa_eig$vectors,
  col = c("red", "green"),
  radius = sa_eig$values, lwd = 2
)
graphics::legend("topright",
  legend = round(sa_eig$values, 2),
  col = c("red", "green"), lty = 1
)

principal_direction(sa.por$azi.PoR)

orientation_strength(sa.por$azi.PoR)
}
\seealso{
\code{\link[=ortensor2d]{ortensor2d()}}
}
