% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ortensor2d}
\alias{ortensor2d}
\title{Orientation tensor}
\usage{
ortensor2d(x, w = NULL, norm = FALSE)
}
\arguments{
\item{x}{numeric vector. Values in degrees.}

\item{w}{(optional) Weights. A vector of positive numbers and of the same
length as \code{x}.}

\item{norm}{logical. Whether the tensor should be normalized.}
}
\value{
2x2 matrix
}
\description{
Orientation tensor
}
\note{
\deqn{E = x \cdot x^{T}}
}
\examples{
test <- rvm(100, mean = 0, k = 10)
ortensor2d(test)
}
\seealso{
\code{\link[=ot_eigen2d]{ot_eigen2d()}}
}
