% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{ot_eigen2d}
\alias{ot_eigen2d}
\title{Decomposition of orientation tensor}
\usage{
ot_eigen2d(x, w = NULL, axial = TRUE, scale = FALSE)
}
\arguments{
\item{x}{numeric vector. Values in degrees.}

\item{w}{(optional) Weights. A vector of positive numbers and of the same
length as \code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{scale}{logical. Whether the Eigenvalues should be scaled so they sum up to 1.}
}
\value{
list of Eigenvalues and the angles corresponding to the Eigenvectors.
}
\description{
Eigenvector decomposition of the orientations.
}
\details{
Eigenvalues can be interpreted as the fraction of the data explained by the
orientation of the associated Eigenvector.
}
\examples{
test <- rvm(100, mean = 0, k = 10)
ot_eigen2d(test, axial = FALSE)

data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
sa_eig <- ot_eigen2d(sa.por$azi.PoR, w = weighting(san_andreas$unc), scale = TRUE)
print(sa_eig)

rose(sa.por$azi.PoR, muci = FALSE)
rose_line(sa_eig$vectors, col = c('red', 'green'),
  radius = sa_eig$values, lwd = 2)
graphics::legend("topright",
  legend = round(sa_eig$values, 2),
  col = c('red', 'green'), lty = 1)
}
\seealso{
\code{\link[=ortensor2d]{ortensor2d()}}
}
