% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{plot_density}
\alias{plot_density}
\title{Circular Kernel Density Plot}
\usage{
plot_density(
  x,
  bw = NULL,
  kernel = c("vonmises", "wrappedcauchy", "wrappednormal"),
  weights = NULL,
  axial = TRUE,
  n = 512L,
  norm.density = TRUE,
  kappa = NULL,
  rho = NULL,
  sd = NULL,
  c = NULL,
  fill = FALSE,
  scale = 0,
  shrink = 1,
  add = TRUE,
  main = NULL,
  labels = TRUE,
  at = seq(0, 360 - 45, 45),
  cborder = TRUE,
  grid = FALSE,
  ...
)
}
\arguments{
\item{x}{Either an object of class \code{"density"} or a numeric vector of angles
(in degrees) from which the estimate is to be computed}

\item{bw, kappa, rho, sd, c}{numeric. Smoothing bandwidth expressed as the concentration
parameter \eqn{\kappa}{k} for the von Mises distribution, \eqn{\rho}{rho} for the
wrapped Cauchy distribution, or \eqn{\sigma}{sd} for the wrapped normal distribution.
Small and large values for the von Mises and wrapped normal/Cauchy distribution,
respectively, gives smooth density lines. If not specified, parameter will be estimated using
\code{\link[=est.kappa]{est.kappa()}}  for the von Mises distribution, or set to \eqn{p \exp(-1)}{p exp(-1)} and \code{1}
for the wrapped Cauchy and wrapped Normal distribution (where \eqn{p = 2}{p=2}
when \code{axial=TRUE} and 1 otherwise), respectively.}

\item{kernel}{character. The smoothing kernel to be used; one of \code{"vonmises"}
(the default), \code{"wrappedcauchy"}, \verb{"wrappednormal}, for the von Mises,
the wrapped Cauchy, and the wrapped Normal distribution.}

\item{weights}{numeric. A vector of observation weights, of the same length as \code{x},
to give individual observations weight in the density estimate. Should sum to 1;
a warning is issued if it doesn't (unless \code{subdensity = TRUE}). Defaults to
equal weight \code{1/length(x)} per observation, matching \code{\link[stats:density]{stats::density()}}.}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})
or biaxial (\code{TRUE}, the default).}

\item{n}{integer. Number of equally spaced angles at which the density is
to be estimated.}

\item{norm.density}{logical. Normalize the density?}

\item{fill}{logical. Whether to fill the density curve or draw just a line
(the default)}

\item{scale}{numeric. radius of plotted circle. Default is \code{1.1}.}

\item{shrink}{numeric. parameter that controls the size of the plotted
function. Default is \code{1}.}

\item{add}{logical. Add to existing plot? (\code{TRUE} by default).}

\item{main}{Character string specifying the title of the plot.}

\item{labels}{Either a logical value indicating whether to plot labels
next to the tick marks, or a vector of labels for the tick marks.}

\item{at}{Optional vector of angles at which tick marks should be plotted.
Set \code{at=numeric(0)} to suppress tick marks.}

\item{cborder}{logical. Border of rose plot.}

\item{grid}{logical. Whether a grid should be added.}

\item{...}{Further graphical parameters may also be supplied as arguments.}
}
\value{
plot or calculated densities
}
\description{
Plots multiples of a von Mises, wrapped Cauchy, and wrapped Normal
density distribution in a circular plot
}
\examples{
# Filled von Mises kernel density curve inside the plot
plot_density(san_andreas$azi,
  kappa = 100,
  fill = TRUE, col = "#51127C80", border = "#51127CFF",
  grid = TRUE,
  add = FALSE
)

# Superimpose a wrapped Cauchy kernel distribution curve
plot_density(san_andreas$azi,
  rho = 0.9, kernel = "wrappedcauchy",
  fill = FALSE, col = "#FB8861FF",
  add = TRUE
)

# Superimpose a wrapped Normal kernel distribution curve
plot_density(san_andreas$azi,
  sd = 2, kernel = "wrappednormal",
  fill = FALSE, col = "#E65164FF",
  add = TRUE
)

# Superimpose a von Mises kernel density curve on a rose diagram:
rose(san_andreas$azi, grid = TRUE)
plot_density(san_andreas$azi,
  bw = 100, col = "#51127CFF",
  add = TRUE, lwd = 3
)

# Corona plot (density curve outside of a rose diagram plot):
w <- weighting(san_andreas$unc)
rose(san_andreas$azi, weights = w, dots = TRUE, stack = TRUE, dot_cex = 0.5, dot_pch = 21)
plot_density(san_andreas$azi, weights = w,
  bw = 100,
  scale = 1.1, shrink = 3, xpd = NA,
  col = "#51127CFF"
)
}
\seealso{
\code{\link[=circular_density]{circular_density()}}

Other circular-plot: 
\code{\link{plot_points}()},
\code{\link{rose}()},
\code{\link{rose_geom}},
\code{\link{rose_stats}()}
}
\concept{circular-plot}
