% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR2Geo_azimuth}
\alias{PoR2Geo_azimuth}
\title{Azimuth Conversion From PoR to Geographical Coordinate Reference System}
\usage{
PoR2Geo_azimuth(x, PoR, axial = TRUE)
}
\arguments{
\item{x}{\code{sf} object or a \code{data.frame} containing the coordinates of the
point(s) (\code{lat}, \code{lon} columns). \code{x} must contain the direction of
\eqn{\sigma_\text{Hmax}}{SHmax} as column \code{azi}, its standard deviation
(column \code{unc}) is optional).}

\item{PoR}{\code{data.frame} or object of class \code{euler.pole} containing the
geographical coordinates of the Eule pole.}

\item{axial}{logical. Whether the azimuth is axial (0-180°) or directional
(0-360°).}
}
\value{
numeric vector of transformed azimuths (in degrees)
}
\description{
Conversion of PoR azimuths into geographical azimuths
}
\examples{
data("nuvel1")
# North America relative to Pacific plate:
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
data("san_andreas")
san_andreas$azi.PoR <- PoR_shmax(san_andreas, PoR)

# convert back to geo CRS
PoR2Geo_azimuth(san_andreas, PoR) |> head()
}
\references{
Stephan, T., Enkelmann, E., and Kroner, U. "Analyzing the
horizontal orientation of the crustal stress adjacent to plate boundaries".
\emph{Sci Rep} 13. 15590 (2023). \doi{10.1038/s41598-023-42433-2}.
}
\seealso{
\code{\link[=PoR_shmax]{PoR_shmax()}}
}
