% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR2geo_shmax}
\alias{PoR2geo_shmax}
\title{Azimuth conversion from PoR to geographical coordinate reference system}
\usage{
PoR2geo_shmax(df, euler)
}
\arguments{
\item{df}{\code{data.frame} containing the PoR-equivalent coordinates of the
point(s) (\code{lat.PoR}, \code{lon.PoR}) and the PoR-equivalent azimuth
(\code{azi.PoR})}

\item{euler}{\code{data.frame} containing the geographical location of
the Euler pole (\code{lat}, \code{lon})}
}
\description{
Helper function to convert PoR azimuths into geographical azimuths
}
\examples{
\dontrun{
data("nuvel1")
# North America relative to Pacific plate:
euler <- subset(nuvel1, nuvel1$plate.rot == "na")
data("san_andreas")
head(san_andreas$azi)
azi.PoR <- PoR_shmax(san_andreas, euler)
res.PoR <- data.frame(azi.PoR = azi.PoR, PoR_coordinates(san_andreas, euler))
res.geo <- PoR2geo_shmax(res.PoR, euler)
head(res.geo)
}
}
