% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{PoR_coordinates}
\alias{PoR_coordinates}
\title{PoR coordinates}
\usage{
PoR_coordinates(x, ep)
}
\arguments{
\item{x}{\code{sf} or \code{data.frame} containing lat and lon coordinates
(\code{lat}, \code{lon})}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole
(\code{lat}, \code{lon})}
}
\value{
\code{data.frame} with the PoR coordinates
(\code{lat.PoR}, \code{lon.PoR})
}
\description{
Retrieve the PoR equivalent coordinates of an object
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- PoR_coordinates(san_andreas, euler)
head(san_andreas.por)
}
