% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{PoR_crs}
\alias{PoR_crs}
\title{PoR coordinate reference system}
\usage{
PoR_crs(x)
}
\arguments{
\item{x}{\code{data.frame} of the geographical coordinates of the Euler pole}
}
\description{
Helper function to create the reference system transformed in Euler pole
coordinate
}
\details{
The PoR coordinate reference system is oblique transformation of the
geographical coordinate system with the Euler pole coordinates being the the
translation factors.
}
