% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_shmax.R
\name{PoR_shmax}
\alias{PoR_shmax}
\title{Theoretical Direction of Maximum Horizontal Stress in PoR reference
system}
\usage{
PoR_shmax(df, euler, type = c("none", "in", "out", "right", "left"))
}
\arguments{
\item{df}{\code{data.frame} containing the coordinates of the point(s)
(\code{lat}, \code{lon}), the direction of
\eqn{\sigma_{Hmax}}{SHmax} \code{azi} and its standard deviation
\code{unc} (optional)}

\item{euler}{\code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler  pole}

\item{type}{Character. Type of plate boundary (optional). Can be
\code{"out"}, \code{"in"}, \code{"right"}, or
\code{"left"} for outward, inward, right-lateral, or left-lateral
moving plate boundaries, respectively. If \code{"none"} (the default), only
the PoR-equivalent azimuth is returned.}
}
\value{
Either a numeric vector of the azimuths in the transformed coordinate
system, or a \code{"data.frame"} with
\describe{
\item{\code{"azi.PoR"}}{the transformed azimuths,}
\item{\code{"prd"}}{the predicted azimuths,}
\item{\code{"dev"}}{the deviation, and}
\item{\code{"nchisq"}}{the normalized \eqn{\chi^2}{chi-squared} statistics.}
}
}
\description{
Models the direction of maximum horizontal stress
\eqn{\sigma_{Hmax}}{SHmax} in the Euler pole (Pole of Rotation)
coordinate reference system. When type of plate boundary is given, it also
gives the deviation from the theoretically predicted azimuth of
\eqn{\sigma_{Hmax}}{SHmax}, the deviation, and the normalized
\eqn{\chi^2}{chi-squared} statistics.
}
\details{
According to the theory, the azimuth of
\eqn{\sigma_{Hmax}}{SHmax} in the pole of rotation reference system is
approximate 0 (or 180), 45, 90, 135 degrees if the stress is sourced by an
outward, sinistral, inward, or dextral moving plate boundary, respectively.
directions of \eqn{\sigma_{Hmax}}{SHmax} with respect to the four
plate boundary types.
}
\examples{
data("nuvel1")
# North America relative to Pacific plate:
euler <- subset(nuvel1, nuvel1$plate.rot == "na")

data("san_andreas")
res <- PoR_shmax(san_andreas, euler, type = "right")
head(res)
}
\seealso{
\code{\link[=model_shmax]{model_shmax()}} to compute the theoretical direction of
\eqn{\sigma_{Hmax}}{SHmax} in the geographical reference system.
\code{\link[=misfit_shmax]{misfit_shmax()}} to compute the deviation of the modeled direction
from the observed direction of \eqn{\sigma_{Hmax}}{SHmax}.
\code{\link[=norm_chisq]{norm_chisq()}} to calculate the normalized \eqn{\chi^2}{chi-squared}
statistics.
}
