% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{PoR_stress2grid}
\alias{PoR_stress2grid}
\title{Stress2Grid in PoR system}
\usage{
PoR_stress2grid(x, ep, ...)
}
\arguments{
\item{x}{\code{sf} object containing
\describe{
\item{azi}{SHmax in degree}
\item{unc}{Uncertainties of SHmax in degree}
\item{type}{Methods used for the determination of the orientation of SHmax}
}}

\item{ep}{Euler pole coordinates}

\item{...}{Arguments passed to \code{\link[=stress2grid]{stress2grid()}}}
}
\description{
The data is transformed into the PoR system before the
interpolation. The interpolation grid is returned in geographical coordinates
and azimuths.
}
\details{
Stress field interpolation in PoR system
}
\examples{
data("san_andreas")
data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
PoR_stress2grid(san_andreas, ep)
}
