% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{PoR_stress2grid}
\alias{PoR_stress2grid}
\title{Spatial interpolation of SHmax in PoR coordinate reference system}
\usage{
PoR_stress2grid(
  x,
  euler,
  grid = NULL,
  PoR_grid = TRUE,
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  ...
)
}
\arguments{
\item{x}{\code{sf} object containing
\describe{
\item{azi}{SHmax in degree}
\item{unc}{Uncertainties of SHmax in degree}
\item{type}{Methods used for the determination of the orientation of SHmax}
}}

\item{euler}{\code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler  pole}

\item{grid}{(optional) Point object of class \code{sf}.}

\item{PoR_grid}{logical. Whether the grid should be generated based on the
coordinate range in the PoR (\verb{"TRUE}, the default) CRS or the geographical CRS
(\verb{"FALSE}). Is ignored if \code{"grid"} is specified.}

\item{lon_range, lat_range}{(optional) numeric vector specifying the minimum
and maximum longitudes and latitudes (are ignored if \code{"grid"} is specified).}

\item{gridsize}{Numeric. Target spacing of the regular grid in decimal
degree. Default is 2.5 (is ignored if \code{"grid"} is specified)}

\item{...}{Arguments passed to \code{\link[=stress2grid]{stress2grid()}}}
}
\description{
The data is transformed into the PoR system before the
interpolation. The interpolation grid is returned in geographical coordinates
and azimuths.
}
\details{
Stress field and wavelength analysis in PoR system and back-transformed
}
\examples{
data("san_andreas")
data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
PoR_stress2grid(san_andreas, ep)
}
\seealso{
\code{\link[=stress2grid]{stress2grid()}}, \code{\link[=compact_grid]{compact_grid()}}
}
