% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_transformation}
\alias{por_transformation}
\alias{PoR_to_geographical}
\alias{geographical_to_PoR}
\title{Conversion between PoR to geographical coordinates}
\usage{
PoR_to_geographical(x, ep)

geographical_to_PoR(x, ep)
}
\arguments{
\item{x}{\code{sf}, \code{SpatRast}, or \code{Raster*} object of the data
points in geographical or PoR coordinate system}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler
pole (\code{lat}, \code{lon})}
}
\value{
\code{sf} or \code{SpatRast} object of the data points in the
transformed geographical or PoR coordinate system
}
\description{
Transform spherical objects from PoR to geographical coordinate reference
system and vice versa.
}
\details{
The PoR coordinate reference system is oblique transformation of the
geographical coordinate system with the Euler pole coordinates being the
translation factors.
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- geographical_to_PoR(san_andreas, euler)
PoR_to_geographical(san_andreas.por, euler)
}
