% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_transformation_df}
\alias{por_transformation_df}
\alias{geographical_to_PoR_df}
\alias{PoR_to_geographical_df}
\title{Conversion between spherical PoR to geographical coordinate system}
\usage{
geographical_to_PoR_df(x, PoR)

PoR_to_geographical_df(x, PoR)
}
\arguments{
\item{x}{\code{"data.frame"} containing \code{lat} and \code{lon}
coordinates of a point in the geographical CRS or the \code{lat.PoR},
\code{lon.PoR}) of the point in the PoR CRS.}

\item{PoR}{Pole of Rotation. \code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler pole}
}
\value{
\code{"data.frame"} with the transformed coordinates
(\code{lat.PoR} and \code{lon.PoR} for PoR CRS,
or \code{lat} and \code{lon} for geographical CRS).
}
\description{
Transformation from spherical PoR to geographical coordinate system and
vice versa
}
