% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_transformation_sf}
\alias{por_transformation_sf}
\alias{PoR_to_geographical_sf}
\alias{geographical_to_PoR_sf}
\title{Conversion between PoR to geographical coordinates of spatial data}
\usage{
PoR_to_geographical_sf(x, euler)

geographical_to_PoR_sf(x, euler)
}
\arguments{
\item{x}{\code{sf}, \code{SpatRast}, or \code{Raster*} object of the data
points in geographical or PoR coordinate system}

\item{euler}{\code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler pole}
}
\value{
\code{sf} or \code{SpatRast} object of the data points in the
transformed geographical or PoR coordinate system
}
\description{
Transform spatial objects from PoR to geographical coordinate reference
system and vice versa.
}
\details{
The PoR coordinate reference system is oblique transformation of the
geographical coordinate system with the Euler pole coordinates being the
translation factors.
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- geographical_to_PoR_sf(san_andreas, euler)
PoR_to_geographical_sf(san_andreas.por, euler)
}
