% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{por_transformation_sf}
\alias{por_transformation_sf}
\alias{PoR_to_geographical_sf}
\alias{geographical_to_PoR_sf}
\title{Conversion between PoR to geographical coordinates of spatial data}
\usage{
PoR_to_geographical_sf(x, PoR)

geographical_to_PoR_sf(x, PoR)
}
\arguments{
\item{x}{\code{sf}, \code{SpatRast}, or \code{Raster*} object of the data
points in geographical or PoR coordinate system}

\item{PoR}{Pole of Rotation. \code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler pole}
}
\value{
\code{sf} or \code{SpatRast} object of the data points in the
transformed geographical or PoR coordinate system
}
\description{
Transform spatial objects from PoR to geographical coordinate reference
system and vice versa.
}
\details{
The PoR coordinate reference system is oblique transformation of the
geographical coordinate system with the Euler pole coordinates being the
translation factors.
}
