% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{PositionCenterSpoke}
\alias{PositionCenterSpoke}
\title{Centrically aligned geom_spoke marker}
\format{
An object of class \code{PositionCenterSpoke} (inherits from \code{Position}, \code{ggproto}, \code{gg}) of length 2.
}
\source{
\url{https://stackoverflow.com/questions/55474143/how-to-center-geom-spoke-around-their-origin}
}
\usage{
PositionCenterSpoke
}
\description{
\code{"position"} subclass "center_spoke" to center
\code{ggplot::geom_spoke()} marker at its origin
}
\keyword{datasets}
