% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{raster_transformation}
\alias{raster_transformation}
\alias{geographical_to_PoR_raster}
\alias{PoR_to_geographical_raster}
\title{Conversion between PoR to geographical coordinate reference system of raster
data}
\usage{
geographical_to_PoR_raster(x, PoR)

PoR_to_geographical_raster(x, PoR)
}
\arguments{
\item{x}{\code{"SpatRaster"} or \code{"RasterLayer"}}

\item{PoR}{Pole of Rotation. \code{"data.frame"} or object of class \code{"euler.pole"}
containing the geographical coordinates of the Euler pole}
}
\value{
terra "SpatRaster" object
}
\description{
Helper function to transform raster data set from PoR to geographical
coordinates
}
