% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{raster_transformation}
\alias{raster_transformation}
\alias{geographical_to_PoR_raster}
\alias{PoR_to_geographical_raster}
\title{Conversion between PoR to geographical coordinate system of raster data}
\usage{
geographical_to_PoR_raster(x, ep)

PoR_to_geographical_raster(x, ep)
}
\arguments{
\item{x}{\code{"SpatRaster"} or \code{"RasterLayer"}}

\item{ep}{\code{data.frame} of the geographical coordinates of the Euler pole
(\code{lat}, \code{lon})}
}
\value{
"SpatRaster"
}
\description{
Helper function to transform raster dataset from PoR to geographical
coordinate system
}
