% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{rayleigh_test}
\alias{rayleigh_test}
\title{Rayleigh Test of Uniformity}
\usage{
rayleigh_test(x, w = NULL, axial = TRUE, na.rm = TRUE, mean_direction = NULL)
}
\arguments{
\item{x}{numeric vector. Values in degrees, for which the
mean, median or standard deviation are required.}

\item{w}{(optional) Weights. A vector of positive numbers, of the same length as
\code{x}.}

\item{axial}{logical. Whether the data are axial, i.e. pi-periodical
(TRUE, the default) or circular, i.e. 2pi-periodical (FALSE).}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}

\item{mean_direction}{optional. The given or known mean direction (in degrees)}
}
\value{
a list with the components:
number of data,
the mean resultant length,
the test score statistic,
the modified test score statistic, and
the significance level (p-value) of the test statistic.
}
\description{
Performs a Rayleigh test of uniformity (or randomness), assessing the
significance of the mean resultant length.
The alternative hypothesis is an unimodal distribution with unknown mean
direction and unknown mean resultant length.
}
\details{
When the test statistic exceeds the significance level, the null hypothesis (concentration parameter is equal to zero) is rejected.
Thus the data have a non-random, preferred trend.
}
\examples{
pidgeon_homing <- c(55,60,65,95,100,110,260,275,285,295)
rayleigh_test(pidgeon_homing, axial = FALSE)

atomic_weight <- c(rep(0, 12), rep(3.6, 1), rep(36, 6), rep(72,1), rep(108,2), rep(169.2, 1), rep(324, 1))
rayleigh_test(atomic_weight, axial = FALSE, mean_direction = 0)

finland_stria <- c(23, 27, 53, 58, 64, 83, 85, 88, 93, 99, 100, 105, 113, 113, 114, 117, 121, 123, 125, 126, 126, 126, 127, 127, 128, 128, 129, 132, 132, 132, 134, 135, 137, 144, 145, 145, 146, 153, 155, 155, 155, 157, 163, 165, 171, 172, 179, 181, 186, 190, 212)
rayleigh_test(finland_stria, axial = FALSE)

data(san_andreas)
data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, ep, "right")
rayleigh_test(san_andreas$azi, 1/san_andreas$unc)
rayleigh_test(sa.por$azi.PoR, 1/san_andreas$unc)
rayleigh_test(sa.por$azi.PoR, mean_direction = 135)
}
\references{
Mardia and Jupp (2000). Directional Statistics. John Wiley and Sons.
}
\seealso{
\code{\link[=mean_resultant_length]{mean_resultant_length()}}, \code{\link[=circular_mean]{circular_mean()}}
}
