% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{roll_circstats}
\alias{roll_circstats}
\title{Apply Rolling Functions using circular statistics}
\usage{
roll_circstats(
  azi,
  w = NULL,
  FUN,
  width,
  by.column = FALSE,
  partial = TRUE,
  align = "center",
  fill = NA,
  ...
)
}
\arguments{
\item{azi}{numeric. Azimuth of \eqn{\sigma_{Hmax}}{SHmax}}

\item{...}{
  Arguments passed on to \code{\link[zoo:rollapply]{zoo::rollapply}}
  \describe{
    \item{\code{}}{}
  }}

\item{unc}{numeric. Uncertainty of observed \eqn{\sigma_{Hmax}}{SHmax}, either a
numeric vector or a number}
}
\value{
numeric vector  with the results of the rolling function.
}
\description{
A generic function for applying a function to rolling margins of an array.
}
\note{
If the rolling statistics are applied to values that are a function of
distance it is recommended to sort the values first.
}
\examples{
data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
x = san_andreas,
euler = ep,
pb = plate_boundary,
tangential = TRUE
)
dat <- san_andreas[order(distance),]
roll_circstats(dat$azi, w = 1/dat$unc, circular_mean, width = 51)
}
