% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{rose}
\alias{rose}
\title{Rose Diagram}
\usage{
rose(
  x,
  bins = NULL,
  binwidth = NULL,
  axial = TRUE,
  clockwise = TRUE,
  unit = c("degree", "radian"),
  main = "N",
  sub,
  ...
)
}
\arguments{
\item{x}{Data to be plotted. A numeric vector containing angles.}

\item{bins}{number of arcs to partition the circle width.
Overridden by \code{binwidth}.}

\item{binwidth}{The width of the bins.}

\item{axial}{Logical. Whether data are uniaxial (axial=FALSE)
or biaxial (axial=TRUE, the default).}

\item{clockwise}{Logical. Whether angles increase in the
clockwise direction (clockwise=TRUE, the default) or anti-clockwise,
counter-clockwise direction (clockwise=FALSE).}

\item{unit}{The unit in which the angles are expressed. "degrees" the default,
or "radians".}

\item{main, sub}{Character string specifying the title and subtitle of the
plot. If \code{sub = NULL}, it will show the bin width.}

\item{...}{Additional arguments passed to \code{spatstat::rose()}.}
}
\value{
A window (class "owin") containing the plotted region.
}
\description{
Plots a rose diagram (rose of directions), the analogue of a
histogram or density plot for angular data.
}
\note{
If bins and binwidth are NULL, an optimal bin width will be calculated using:
\deqn{ \frac{2 IQR(x)}{n^{\frac{1}{3}}}
}
with n being the length of x.
}
\examples{
x <- runif(100, 60, 210)
rose(x)

data("san_andreas")
rose(san_andreas$azi, col = "grey", axial = TRUE)
}
