% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rose}
\alias{rose}
\title{Rose Diagram}
\usage{
rose(
  x,
  binwidth = NULL,
  bins = NULL,
  axial = TRUE,
  clockwise = TRUE,
  unit = c("degree", "radian"),
  main = "N",
  sub,
  ...
)
}
\arguments{
\item{x}{Data to be plotted. A numeric vector containing angles.}

\item{binwidth}{The width of the bins.}

\item{bins}{number of arcs to partition the circle width.
Overridden by \code{binwidth}.}

\item{axial}{Logical. Whether data are uniaxial (\code{axial=FALSE})
or biaxial (\code{TRUE}, the default).}

\item{clockwise}{Logical. Whether angles increase in the
clockwise direction (\code{clockwise=TRUE}, the default) or anti-clockwise,
counter-clockwise direction (\code{FALSE}).}

\item{unit}{The unit in which the angles are expressed.
\code{"degrees"} (the default), or \code{"radians"}.}

\item{main, sub}{Character string specifying the title and subtitle of the
plot. If \code{sub = NULL}, it will show the bin width.}

\item{...}{Additional arguments passed to \code{\link[spatstat.explore:rose]{spatstat.explore::rose()}}.}
}
\value{
A window (class \code{"owin"}) containing the plotted region.
}
\description{
Plots a rose diagram (rose of directions), the analogue of a
histogram or density plot for angular data.
}
\note{
If \code{bins} and \code{binwidth} are \code{NULL}, an optimal bin width will be
calculated using:
\deqn{ \frac{2 IQR(x)}{n^{\frac{1}{3}}}
}
with n being the length of \code{x}.
}
\examples{
x <- runif(100, 60, 210)
rose(x)

data("san_andreas")
rose(san_andreas$azi, col = "grey", axial = TRUE)
}
