% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{rotate_lines}
\alias{rotate_lines}
\title{Rotate Lines}
\usage{
rotate_lines(theta, p, centre)
}
\arguments{
\item{theta}{Angle of rotation (in degree)}

\item{p}{Coordinates of the lines end points}

\item{centre}{Coordinates of the center point of rotation}
}
\value{
\code{matrix}
}
\description{
Rotates a set of straight lines around an angle
}
\keyword{internal}
