% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation Matrix}
\usage{
rotation_matrix(n, alpha)
}
\arguments{
\item{n}{Rotation axis (in Cartesian coordinates). Can be a vector of three
numbers or a matrix of 3 columns (x, y, z)}

\item{alpha}{Rotation angle in degrees}
}
\value{
\code{matrix}
}
\description{
Calculates the rotation matrix using the rotation axis and the
angle of rotation
}
\examples{
w <- c(0, 1, 0)
rotation_matrix(w, 90)
}
