% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{smallcircle_dummy}
\alias{smallcircle_dummy}
\title{Small circle grid dummy}
\usage{
smallcircle_dummy(x = 10)
}
\arguments{
\item{x}{angle between small circles. Default is 10}
}
\value{
data.frame
}
\description{
create a dummy
}
\examples{
smallcircle_dummy(30)
smallcircle_dummy()
}
