% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{spec_atan}
\alias{spec_atan}
\alias{atan_spec}
\alias{atand_spec}
\title{Quadrant-specific inverse of the tangent}
\usage{
atan_spec(x, y)

atand_spec(x, y)
}
\arguments{
\item{x, y}{dividend and divisor that comprise the sum of sines and cosines, respectively.}
}
\description{
returns the quadrant specific inverse of the tangent
}
\references{
Jammalamadaka, S. Rao, and Ambar Sengupta. Topics in circular statistics. Vol. 5. world scientific, 2001.
}
