% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{spherical_angle}
\alias{spherical_angle}
\alias{orthodrome_haversine}
\alias{orthodrome_haversine2}
\alias{orthodrome_vincenty}
\title{Angle along great circle on spherical surface}
\usage{
orthodrome_haversine(lat1, lon1, lat2, lon2)

orthodrome_haversine2(lat1, lon1, lat2, lon2)

orthodrome_vincenty(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1, lat2}{numeric vector; latitudes of point 1 and 2 (in radians)}

\item{lon1, lon2}{numeric vector; longitudes of point 1 and 2 (in radians)}
}
\value{
angle in radians
}
\description{
Smallest angle between two points on the surface of a sphere, measured along
the surface of the sphere
}
\details{
\describe{
\item{"orthodrome_haversine"}{uses Haversine formula (the default)}
\item{"orthodrome_haversine2"}{ uses Haversine formula optimized for 64-bit
floating-point numbers}
\item{"orthodrome_vincenty"}{uses Vincenty formula for an ellipsoid with
equal major and minor axes}
}
}
\examples{
berlin <- c(52.52, 13.41)
calgary <- c(51.04, -114.072)
orthodrome_haversine(berlin[1], berlin[2], calgary[1], calgary[2])
orthodrome_haversine2(berlin[1], berlin[2], calgary[1], calgary[2])
orthodrome_vincenty(berlin[1], berlin[2], calgary[1], calgary[2])
}
