% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{stress2grid}
\alias{stress2grid}
\title{Stress2Grid}
\source{
\url{https://github.com/MorZieg/Stress2Grid}
}
\usage{
stress2grid(
  x,
  stat = c("mean", "median"),
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  min_data = 3,
  threshold = 25,
  arte_thres = 200,
  method_weighting = FALSE,
  quality_weighting = TRUE,
  dist_weight = c("inverse", "linear", "none"),
  dist_threshold = 0.1,
  R_range = seq(50, 1000, 50)
)
}
\arguments{
\item{x}{\code{sf} object containing
\describe{
\item{azi}{SHmax in degree}
\item{unc}{Uncertainties of SHmax in degree}
\item{type}{Methods used for the determination of the orientation of SHmax}
}}

\item{stat}{Should the orientation of interpolated SHmax be based  on the
circular mean and standard deviation (\code{"mean"}, the default) or on the
circular median
and interquartile range (\code{"median"})?}

\item{lon_range, lat_range}{(optional) numeric vector specifying the minimum
and maximum longitudes and latitudes.}

\item{gridsize}{Numeric. Target spacing of the regular grid in decimal degree.
Default is 2.5}

\item{min_data}{Integer. Minimum number of data per bin. Default is 3}

\item{threshold}{Numeric. Threshold for deviation of orientation. Default is
25}

\item{arte_thres}{Numeric. Maximum distance (in km) of the gridpoint to the
next
datapoint. Default is 200}

\item{method_weighting}{Logical. If a method weighting should be applied:
Default is \code{FALSE}.}

\item{quality_weighting}{Logical. If a quality weighting should be applied:
Default is
\code{TRUE}.}

\item{dist_weight}{Distance weighting method which should be used. One of
"none", "linear", or "inverse" (the default).}

\item{dist_threshold}{Numeric. Distance weight to prevent overweight of data
nearby
(0 to 1). Default is 0.1}

\item{R_range}{Numeric value or vector specifying the search radius (im km).
Default is \code{seq(50, 1000, 50)}}
}
\value{
\code{sf} object containing
\describe{
\item{lon,lat}{longitude and latitude in degree}
\item{azi}{Mean SHmax in degree}
\item{sd}{Standard deviation of SHmax in degree}
\item{R}{Search radius in km}
\item{mdr}{Mean distance of datapoint per search radius}
\item{N}{Number of data points in search radius}
}
}
\description{
Stress pattern and wavelength analysis
}
\details{
Calculates the weighted mean and standard deviation of the stress
data
}
\examples{
data("san_andreas")
stress2grid(san_andreas)
}
\references{
Ziegler, M. O. and Heidbach, O. (2019).
Matlab Script Stress2Grid v1.1. GFZ Data Services. \doi{10.5880/wsm.2019.002}
}
