% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_eulerpole.R
\name{stress_paths}
\alias{stress_paths}
\alias{eulerpole_paths}
\alias{eulerpole_smallcircles}
\alias{eulerpole_greatcircles}
\alias{eulerpole_loxodromes}
\title{Theoretical Plate Tectonic Stress Paths}
\usage{
eulerpole_paths(x, type, n = 10, angle, sense, returnclass = c("sf", "sp"))

eulerpole_smallcircles(x, n = 10, returnclass = c("sf", "sp"))

eulerpole_greatcircles(x, n = 10, returnclass = c("sf", "sp"))

eulerpole_loxodromes(x, n = 10, angle = 45, sense, returnclass = c("sf", "sp"))
}
\arguments{
\item{x}{\code{data.frame} containing coordinates of Euler pole in lat, lon,
and rotation angle (optional)}

\item{type}{Character string specifying the type of curves to export. Either
\code{"sm"} for small circles (default), \code{"gc"} for great circles, or
\code{"ld"} for loxodromes.}

\item{n}{Number of equally spaced curves}

\item{angle}{Direction of loxodromes; default = 45}

\item{sense}{Sense of loxodromes  'sinistral' or 'dextral' for 'clockwise'
or 'counterclockwise' loxodromes, respectively}

\item{returnclass}{"sf" (default) for simple features or "sp" for spatial objects}
}
\value{
\code{sf} or \code{SpatialLinesDataFrame}
}
\description{
Construct \eqn{\sigma_\text{Hmax}}{SHmax} lines that are
following small circles, great circles, or loxodromes of an Euler pole for
the relative plate motion.
}
\details{
Maximum horizontal stress can be aligned to three types of curves
related to relative plate motion:
\describe{
\item{Small circles}{Lines that have a constant distance to the Euler pole.
If x contains \code{angle}, output additionally gives absolute
velocity on small circle (degree/Myr -> km/Myr).}
\item{Great circles}{Paths of the the shortest distance between the Euler
pole and its antipodal position.}
\item{Loxodromes}{Lines of constant bearing, i.e. curves cutting small
circles at a constant angle.}
}
}
\examples{
data("nuvel1")
euler <- subset(nuvel1, nuvel1$ID == "na") # North America relative to Pacific plate
euler$angle <- euler$rate
eulerpole_smallcircles(euler)
eulerpole_greatcircles(euler)
eulerpole_loxodromes(x = euler, angle = 45, n = 10, sense = "sinistral")
eulerpole_loxodromes(x = euler, angle = 30, sense = "dextral")
}
\author{
Tobias Stephan
}
