% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectonicr-package.R
\docType{package}
\name{tectonicr}
\alias{tectonicr}
\alias{_PACKAGE}
\alias{tectonicr-package}
\title{library(tectonicr)}
\description{
Modeling the Direction of the Maximum Horizontal Stress using Relative Plate
Motion
}
\details{
Further details and theoretical background are provided by Wdowinski
(1999) and Stephan et al. (under review).
}
\note{
A list of documented functions may be viewed by typing
\code{help(package="tectonicr")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/tectonicr/}
  \item Report bugs at \url{https://github.com/tobiste/tectonicr/issues}
}

}
\author{
Tobias Stephan
}
\keyword{internal}
