% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectonicr-package.R
\docType{package}
\name{tectonicr}
\alias{tectonicr}
\alias{_PACKAGE}
\alias{tectonicr-package}
\title{library(tectonicr)}
\description{
Modeling the Direction of the Maximum Horizontal Stress using Relative Plate
Motion
}
\details{
Further details and theoretical background are provided by Wdowinski
(1999) and Stephan et al. (2023).
}
\note{
A list of documented functions may be viewed by typing
\code{help(package="tectonicr")}.
}
\references{
Wdowinski (1998) "A theory of intraplate tectonics".
\emph{Journal of Geophysical Research: Solid Earth}, 103(3), 5037–5059.

Stephan, T., Kroner, U., and Enkelmann, E. (2023) "Analyzing the horizontal
orientation of the crustal stress adjacent to plate boundaries".
\emph{Scientific Reports}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tobiste.github.io/tectonicr/}
  \item Report bugs at \url{https://github.com/tobiste/tectonicr/issues}
}

}
\author{
Tobias Stephan
}
\keyword{internal}
