% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{transform_matrices}
\alias{transform_matrices}
\alias{rotmat_x}
\alias{rotmat_y}
\alias{rotmat_z}
\title{Basic rotation matrices}
\usage{
rotmat_x(x)

rotmat_y(x)

rotmat_z(x)
}
\arguments{
\item{x}{angle (in degree)}
}
\value{
matrix
}
\description{
Helper functions for constructing the matrices for rotations about the
x-, y-, and z angles with the angle x
}
