% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{trigon}
\alias{trigon}
\alias{sind}
\alias{cosd}
\alias{tand}
\alias{asind}
\alias{acosd}
\alias{atand}
\alias{atan2d}
\title{Trigonometric Functions in Degrees}
\source{
\code{\link[pracma:sind]{pracma::sind()}}, \code{\link[pracma:sind]{pracma::cosd()}}, \code{\link[pracma:sind]{pracma::tand()}},
\code{\link[pracma:sind]{pracma::asind()}}, \code{\link[pracma:sind]{pracma::acosd()}}, \code{\link[pracma:sind]{pracma::atand()}}, and
\code{\link[pracma:sind]{pracma::atan2d()}} from the package "pracma".
}
\usage{
sind(x)

cosd(x)

tand(x)

asind(x)

acosd(x)

atand(x)

atan2d(x1, x2)
}
\arguments{
\item{x, x1, x2}{Numeric or complex vectors.}
}
\value{
Returns a scalar or vector of numeric values.
}
\description{
Trigonometric functions expecting input in degrees, not radians.
}
\examples{
x <- seq(-3, 7, by = 1 / 8)
tx <- cbind(x, cos(pi * x), cospi(x), sin(pi * x), sinpi(x),
  tan(pi * x), tanpi(x),
  deparse.level = 2
)
op <- options(digits = 4, width = 90) # for nice formatting
head(tx)
tx[(x \%\% 1) \%in\% c(0, 0.5), ]
options(op)
}
