% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigonometry_degrees.R
\name{trigon}
\alias{trigon}
\alias{sind}
\alias{cosd}
\alias{tand}
\alias{asind}
\alias{acosd}
\alias{atand}
\alias{atan2d}
\title{Trigonometric Functions in Degrees}
\usage{
sind(x)

cosd(x)

tand(x)

asind(x)

acosd(x)

atand(x)

atan2d(x1, x2)
}
\arguments{
\item{x, x1, x2}{numeric or complex vectors.}
}
\value{
Returns a scalar or vector of numeric values.
}
\description{
Trigonometric functions expecting input in degrees, not radians.
}
\examples{
x <- seq(-3, 7, by = 1/8)
tx <- cbind(x, cos(pi*x), cospi(x), sin(pi*x), sinpi(x),
            tan(pi*x), tanpi(x), deparse.level=2)
op <- options(digits = 4, width = 90) # for nice formatting
head(tx)
tx[ (x \%\% 1) \%in\% c(0, 0.5) ,]
options(op)
}
