% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{vonmises}
\alias{vonmises}
\alias{rvm}
\alias{dvm}
\alias{pvm}
\alias{qvm}
\title{The von Mises Distribution}
\usage{
rvm(n, mean, kappa)

dvm(theta, mean, kappa, log = FALSE, axial = FALSE)

pvm(theta, mean, kappa, from = NULL, tol = 1e-20)

qvm(p, mean = 0, kappa, from = NULL, tol = .Machine$double.eps^(0.6), ...)
}
\arguments{
\item{n}{integer. Number of observations in degrees}

\item{mean}{numeric. The mean vector in degrees.}

\item{kappa}{numeric. Concentration parameter in the range (0, Inf]}

\item{theta}{numeric. Angular value in degrees}

\item{log}{logical. If \code{TRUE}, probabilities p are given as log(p).}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or directional, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{from}{if \code{NULL} is set to \eqn{\mu-\pi}{mu-pi}. This is the value from
which the pvm and qvm are evaluated. in degrees.}

\item{tol}{numeric. The precision in evaluating the distribution function or the quantile.}

\item{p}{numeric. Vector of probabilities with values in \eqn{[0,1]}{[0,1]}.}

\item{...}{parameters passed to \code{\link[stats:integrate]{stats::integrate()}}.}
}
\value{
\code{dvm} gives the density,
\code{pvm} gives the probability of the von Mises distribution function,
\code{rvm} generates random deviates (in degrees), and
\code{qvm} provides quantiles (in degrees).
}
\description{
Density, probability distribution function, quantiles, and random generation
for the circular normal distribution with mean \eqn{\mu} and kappa \eqn{\kappa}.
}
\examples{
set.seed(1)
x <- rvm(5, mean = 90, kappa = 2)

dvm(x, mean = 90, kappa = 2)
dvm(x, mean = 90, kappa = 2, axial = TRUE)

pvm(x, mean = 90, kappa = 2)
qvm(c(.25, .5, .75), mean = 90, kappa = 2)
}
\seealso{
\link{wcauchy} and \link{cunif}
}
