% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{vonmises}
\alias{vonmises}
\alias{rvm}
\alias{dvm}
\alias{pvm}
\title{The von Mises Distribution}
\usage{
rvm(n, mean, kappa)

dvm(theta, mean, kappa)

pvm(theta, mean, kappa)
}
\arguments{
\item{n}{number of observations in degrees}

\item{mean}{mean in degrees}

\item{kappa}{concentration parameter}

\item{theta}{angular value in degrees}
}
\description{
Density, distribution function, and random generation for the circular normal
distribution with mean and kappa.
}
\examples{
x <- rvm(100, mean = 90, k = 100)
dvm(x, mean = 90, k = 100)
}
