% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{watson_test}
\alias{watson_test}
\title{Watson's U2 Test for Circular Uniformity}
\usage{
watson_test(x, alpha = NULL)
}
\arguments{
\item{x}{numeric vector containing the circular data which are expressed in degrees}

\item{alpha}{Significance level of the test. Valid levels are 0.01, 0.05, 0.1.
This argument may be omitted (\code{NULL}, the default), in which case, a range for the p-value will be returned.}
}
\value{
list containing the test statistic \code{statistic} and the significance
level \code{p.value}.
}
\description{
Watson's test for circular uniform distribution.
}
\details{
If \code{statistic > p.value}, the null hypothesis is rejected.
If not, randomness (uniform distribution) cannot be excluded.
}
\examples{
# Example data from Mardia and Jupp (2001), pp. 93
pidgeon_homing <- c(55, 60, 65, 95, 100, 110, 260, 275, 285, 295)
#
data(san_andreas)
data("nuvel1")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, ep, "right")
watson_test(sa.por$azi.PoR, alpha = .05)
}
\references{
Mardia and Jupp (2000). Directional Statistics. John Wiley and Sons.
}
