% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_tests.R
\name{watson_two_sample}
\alias{watson_two_sample}
\alias{watson_two_test}
\alias{watson_two_test_perm}
\title{Watson's Two-Sample Test of Homogeneity}
\usage{
watson_two_test(x, y, alpha = NULL, axial = TRUE, quiet = FALSE)

watson_two_test_perm(x, y, axial = TRUE, n_perm = 1000L)
}
\arguments{
\item{x, y}{numeric vectors. Angles in degrees}

\item{alpha}{Significance level of the test. Valid levels are \code{0.01}, \code{0.05},
and \code{0.1}.
This argument may be omitted (\code{NULL}, the default), in which case, a range
for the p-value will be returned.}

\item{axial}{logical. Whether the data are axial, i.e. \eqn{\pi}-periodical
(\code{TRUE}, the default) or circular, i.e. \eqn{2 \pi}-periodical (\code{FALSE}).}

\item{quiet}{logical. Prints the test's decision.}

\item{n_perm}{integer. Number of permutations}
}
\value{
list
}
\description{
Performs Watson's test for homogeneity on two samples of circular data.
\code{watson_two_test_perm()} uses permutation to estimate p-values.
}
\details{
Watson's two-sample test of homogeneity is performed, and the
results are printed. If alpha is specified and non-zero, the test statistic
is printed along with the critical value and decision. If alpha is omitted,
the test statistic is printed and a range for the p-value of the test is given.

Critical values for the test statistic are obtained using the asymptotic
distribution of the test statistic. It is recommended to use the obtained
critical values and ranges for p-values only for combined sample sizes in
excess of 17. Tables are available for smaller sample sizes and can be found
in Mardia (1972) for instance.
}
\examples{
set.seed(20250411)
x1 <- c(35, 45, 50, 55, 60, 70, 85, 95, 105, 120)
x2 <- c(75, 80, 90, 100, 110, 130, 135, 140, 150, 160, 165)
watson_two_test(x1, x2, axial = FALSE)
watson_two_test_perm(x1, x2, axial = FALSE)

data1 <- rvm(n=20, mean = 0, kappa=3)
data2 <- rvm(n=20, mean = 90, kappa=2)
watson_two_test(data1, data2, axial = FALSE)
watson_two_test_perm(data1, data2, axial = FALSE)


# San Andreas Fault Data:
data(san_andreas)
data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
sa.por <- PoR_shmax(san_andreas, PoR, "right")
watson_two_test(sa.por$azi, 135, alpha = 0.05)
watson_two_test_perm(sa.por$azi, rvm(100, 135, 10))
}
\seealso{
Other Tests: 
\code{\link{ar_test}()},
\code{\link{kuiper_test}()},
\code{\link{norm_chisq}()},
\code{\link{rayleigh-test}},
\code{\link{watson_test}()},
\code{\link{watson_wheeler_test_perm}()},
\code{\link{weighted-rayleigh-test}}
}
\concept{Tests}
