% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{weighted_circle_stats}
\alias{weighted_circle_stats}
\alias{circular_weighted_mean}
\alias{circular_weighted_sd}
\alias{circular_weighted_median}
\alias{circular_weighted_quantiles}
\alias{circular_weighted_IQR}
\title{Weighted Mean and Statistics on Pi-periodic Data}
\usage{
circular_weighted_mean(x, w = NULL, na.rm = TRUE)

circular_weighted_sd(x, w = NULL, na.rm = TRUE)

circular_weighted_median(x, w = NULL, na.rm = TRUE)

circular_weighted_quantiles(x, w = NULL, na.rm = TRUE)

circular_weighted_IQR(x, w = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{Data values. A vector of numeric values in degrees, for which the
mean, median or standard deviation are required.}

\item{w}{Weights. A vector of nonnegative numbers, of the same length as
\code{x}.}

\item{na.rm}{logical value indicating whether \code{NA} values in \code{x}
should be stripped before the computation proceeds.}
}
\description{
Calculate the weighted median and standard deviation
of orientation data. Weighting is based on the reciprocal of the data
uncertainties.
}
\examples{
x <- c(175, 179, 2, 4)
unc <- c(5, 1, 2, 4)
circular_weighted_mean(x, 1 / unc)
circular_weighted_sd(x, 1 / unc)
circular_weighted_median(x, 1 / unc)
circular_weighted_IQR(x, 1/unc)
}
\references{
\itemize{
\item Mardia, K.V. (1972). Statistics of Directional Data: Probability and
Mathematical Statistics. London: Academic Press.
\item Ziegler, M. O.; Heidbach O. (2019). Manual of the Matlab Script
Stress2Grid v1.1. (WSM Technical Report; 19-02),
GFZ German Research Centre for Geosciences. \doi{10.2312/wsm.2019.002}
\item Heidbach, O., Tingay, M., Barth, A., Reinecker, J., Kurfeß, D., & Müller,
B. (2010). Global crustal stress pattern based on the World Stress Map
database release 2008. \emph{Tectonophysics} 482, 3–-15,
\doi{10.1016/j.tecto.2009.07.023}
}
}
